bondy (bondy v1.0.0-beta.16)

Types

id/0

Specs

id() :: 0..9007199254740992.

maybe/1

Specs

maybe(T) :: T | undefined.

uri/0

Specs

uri() :: binary().

wamp_abort/0

Specs

wamp_abort() :: #abort{}.

wamp_authenticate/0

Specs

wamp_authenticate() :: #authenticate{}.

wamp_call/0

Specs

wamp_call() :: #call{}.

wamp_cancel/0

Specs

wamp_cancel() :: #cancel{}.

wamp_challenge/0

Specs

wamp_challenge() :: #challenge{}.

wamp_error/0

Specs

wamp_error() :: #error{}.

wamp_error_map/0

Specs

wamp_error_map() ::
    #{error_uri => uri(), details => map(), args => list(), kwargs => map(), payload => binary()}.

wamp_event/0

Specs

wamp_event() :: #event{}.

wamp_event_received/0

Specs

wamp_event_received() :: #event_received{}.

wamp_goodbye/0

Specs

wamp_goodbye() :: #goodbye{}.

wamp_hello/0

Specs

wamp_hello() :: #hello{}.

wamp_interrupt/0

Specs

wamp_interrupt() :: #interrupt{}.

wamp_invocation/0

Specs

wamp_invocation() :: #invocation{}.

wamp_message/0

Specs

wamp_message() ::
    wamp_hello() |
    wamp_challenge() |
    wamp_authenticate() |
    wamp_welcome() |
    wamp_abort() |
    wamp_goodbye() |
    wamp_error() |
    wamp_publish() |
    wamp_published() |
    wamp_subscribe() |
    wamp_subscribed() |
    wamp_unsubscribe() |
    wamp_unsubscribed() |
    wamp_event() |
    wamp_event_received() |
    wamp_subscriber_received() |
    wamp_call() |
    wamp_cancel() |
    wamp_result() |
    wamp_register() |
    wamp_registered() |
    wamp_unregister() |
    wamp_unregistered() |
    wamp_invocation() |
    wamp_interrupt() |
    wamp_yield().

wamp_publish/0

Specs

wamp_publish() :: #publish{}.

wamp_published/0

Specs

wamp_published() :: #published{}.

wamp_register/0

Specs

wamp_register() :: #register{}.

wamp_registered/0

Specs

wamp_registered() :: #registered{}.

wamp_result/0

Specs

wamp_result() :: #result{}.

wamp_subscribe/0

Specs

wamp_subscribe() :: #subscribe{}.

wamp_subscribed/0

Specs

wamp_subscribed() :: #subscribed{}.

wamp_subscriber_received/0

Specs

wamp_subscriber_received() :: #subscriber_received{}.

wamp_unregister/0

Specs

wamp_unregister() :: #unregister{}.

wamp_unregistered/0

Specs

wamp_unregistered() :: #unregistered{}.

wamp_unsubscribe/0

Specs

wamp_unsubscribe() :: #unsubscribe{}.

wamp_unsubscribed/0

Specs

wamp_unsubscribed() :: #unsubscribed{}.

wamp_welcome/0

Specs

wamp_welcome() :: #welcome{}.

wamp_yield/0

Specs

wamp_yield() :: #yield{}.

Functions

aae_exchanges()

ack(Pid, Ref)

Specs

ack(pid(), reference()) -> ok.

Acknowledges the reception of a WAMP message. This function should be used by the peer transport module to acknowledge the reception of a message sent with send/3.

add_via(Relay, Opts)

call(Uri, Opts, Args, KWArgs, Ctxt0)

Specs

call(binary(), map(), list() | undefined, map() | undefined, bondy_context:t()) ->
        {ok, map()} | {error, wamp_error_map()}.

A blocking call.

cast(ProcedureUri, Opts, Args, KWArgs, Ctxt0)

Specs

cast(binary(), map(), list() | undefined, map() | undefined, bondy_context:t()) ->
        ok | {error, wamp_error_map()}.

A non-blocking call.

check_response(Uri, ReqId, Timeout, Ctxt)

peek_via(_)

Returns the last relay reference of the 'via' option stack. This reference represents the final relay the message will need to go through to be send using send/3.

prepare_send(To, Opts)

Specs

prepare_send(To :: bondy_ref:t(), Opts :: map()) -> {bondy_ref:t(), map()}.

prepare_send(To, Origin, Opts)

Specs

prepare_send(To :: bondy_ref:t(),
             Origin :: maybe(bondy_ref:client() | bondy_ref:internal()),
             Opts :: map()) ->
                {bondy_ref:t(), map()}.

publish(Opts, TopicUri, Args, KWArgs, CtxtOrRealm)

publish(ReqId, Opts, TopicUri, Args, KWArgs, CtxtOrRealm)

request(Pid, RealmUri, M)

Specs

request(Pid :: pid(), RealmUri :: uri(), M :: wamp_message:t()) -> tuple().

send(Ref, M)

Specs

send(bondy_ref:t(), wamp_message()) -> ok.

Sends a message to a WAMP peer. It calls send/3 with a an empty map for Options.

send(Ref, Msg, Opts)

Specs

send(Ref :: bondy_ref:t(), Msg :: wamp_message(), Opts :: map()) -> ok | no_return().

Sends a message to a local WAMP peer. If the transport is not open it fails with an exception. This function is used by the router (dealer | broker) to send WAMP messages to local peers. Opts is a map with the following keys:
* timeout - timeout in milliseconds (defaults to 10000) * enqueue (boolean) - if the peer is not reachable and this value is true, bondy will enqueue the message so that the peer can resume the session and consume all enqueued messages.

subscribe(RealmUri, Opts, TopicUri)

Calls bondy_broker:subscribe/3.

subscribe(RealmUri, Opts, TopicUri, Fun)

Calls bondy_broker:subscribe/4.

take_via(Opts)

Removes and returns the first relay reference of the 'via' option stack.