bondy_auth behaviour (bondy v1.0.0-beta.16)

This module provides the behaviour to be implemented by the authentication methods used by Bondy. The module provides the functions required to setup an authentication context, compute a challenge (in the case of challenge-response methods and authenticate a user based on the selected method out of the available methods offered by the Realm and restricted by the access control system and the user's password capabilities.

Types

context/0

Specs

context() ::
    #{session_id := id() | undefined,
      realm_uri := uri(),
      user := bondy_rbac_user:t() | undefined,
      user_id := binary() | undefined,
      available_methods := [binary()],
      role := binary(),
      roles := [binary()],
      conn_ip := [{ip, inet:ip_address()}],
      provider => binary(),
      method => binary(),
      callback_mod => module(),
      callback_mod_state => term()}.

id/0

Specs

id() :: 0..9007199254740992.

requirements/0

Specs

requirements() ::
    #{identification := boolean,
      password := {true, #{protocols := [cra | scram]}} | boolean(),
      authorized_keys := boolean(),
      any => requirements(),
      all => requirements()}.

uri/0

Specs

uri() :: binary().

Callbacks

authenticate/4

Specs

authenticate(Signature :: binary(), DataIn :: map(), Ctxt :: context(), CBState :: term()) ->
                {ok, DataOut :: map(), CBState :: term()} |
                {error, Reason :: any(), CBState :: term()}.

challenge/3

Specs

challenge(DataIn :: map(), Ctxt :: context(), CBState :: term()) ->
             {ok, CBState :: term()} |
             {ok, ChallengeData :: map(), CBState :: term()} |
             {error, Reason :: any(), CBState :: term()}.

init/1

Specs

init(Ctxt :: context()) -> {ok, CBState :: term()} | {error, Reason :: any()}.

requirements/0

Specs

requirements() -> requirements().

Functions

authenticate(Method, Signature, DataIn, Ctxt)

Specs

authenticate(Method :: binary(), Signature :: binary(), DataIn :: map(), Ctxt :: context()) ->
                {ok, ReturnExtra :: map(), NewCtxt :: context()} | {error, Reason :: any()}.

available_methods(_)

Specs

available_methods(context()) -> [binary()].

available_methods(List, Ctxt)

Specs

available_methods(List :: [binary()], Ctxt :: context()) -> [binary()].

Returns the sublist of List containing only the available authentication methods that can be used with user User in realm Realm when connecting from the current IP Address.

challenge(Method, DataIn, Ctxt)

Specs

challenge(Method :: binary(), DataIn :: map(), Ctxt :: context()) ->
             {ok, AuthData :: map(), NewCtxt :: context()} |
             {challenge, ChallengeData :: map(), NewCtxt :: context()} |
             {error, Reason :: any()}.

conn_ip(_)

Specs

conn_ip(context()) -> [{ip, inet:ip_address()}].

format_status(Opt, Ctxt)

Specs

format_status(Opt :: normal | terminate, Ctxt :: context()) -> term().

init(SessionId, Realm, UserId, Roles, Peer)

Specs

init(SessionId :: id(),
     Realm :: bondy_realm:t() | uri(),
     UserId :: binary() | anonymous,
     Roles :: all | binary() | [binary()] | undefined,
     Peer :: {inet:ip_address(), inet:port_number()}) ->
        {ok, context()} |
        {error, {no_such_user, binary()} | no_such_realm | no_such_group} |
        no_return().

method(_)

Specs

method(context()) -> [binary()].

method_info()

Specs

method_info() -> map().

method_info(Method)

Specs

method_info(Method :: binary()) -> map() | no_return().

methods()

Specs

methods() -> [binary()].

provider(_)

Specs

provider(context()) -> [binary()].

realm_uri(_)

Specs

realm_uri(context()) -> uri().

role(_)

Specs

role(context()) -> binary().

roles(_)

Specs

roles(context()) -> [binary()].

session_id(_)

Specs

session_id(context()) -> id().

user(_)

Specs

user(context()) -> bondy_rbac_user:t() | undefined.

user_id(_)

Specs

user_id(context()) -> binary() | undefined.