bondy_broker (bondy v1.0.0-beta.16)

This module implements the capabilities of a Broker. It is used by bondy_router.

Types

id/0

Specs

id() :: 0..9007199254740992.

uri/0

Specs

uri() :: binary().

wamp_abort/0

Specs

wamp_abort() :: #abort{}.

wamp_authenticate/0

Specs

wamp_authenticate() :: #authenticate{}.

wamp_call/0

Specs

wamp_call() :: #call{}.

wamp_cancel/0

Specs

wamp_cancel() :: #cancel{}.

wamp_challenge/0

Specs

wamp_challenge() :: #challenge{}.

wamp_error/0

Specs

wamp_error() :: #error{}.

wamp_event/0

Specs

wamp_event() :: #event{}.

wamp_event_received/0

Specs

wamp_event_received() :: #event_received{}.

wamp_goodbye/0

Specs

wamp_goodbye() :: #goodbye{}.

wamp_hello/0

Specs

wamp_hello() :: #hello{}.

wamp_interrupt/0

Specs

wamp_interrupt() :: #interrupt{}.

wamp_invocation/0

Specs

wamp_invocation() :: #invocation{}.

wamp_message/0

Specs

wamp_message() ::
    wamp_hello() |
    wamp_challenge() |
    wamp_authenticate() |
    wamp_welcome() |
    wamp_abort() |
    wamp_goodbye() |
    wamp_error() |
    wamp_publish() |
    wamp_published() |
    wamp_subscribe() |
    wamp_subscribed() |
    wamp_unsubscribe() |
    wamp_unsubscribed() |
    wamp_event() |
    wamp_event_received() |
    wamp_subscriber_received() |
    wamp_call() |
    wamp_cancel() |
    wamp_result() |
    wamp_register() |
    wamp_registered() |
    wamp_unregister() |
    wamp_unregistered() |
    wamp_invocation() |
    wamp_interrupt() |
    wamp_yield().

wamp_publish/0

Specs

wamp_publish() :: #publish{}.

wamp_published/0

Specs

wamp_published() :: #published{}.

wamp_register/0

Specs

wamp_register() :: #register{}.

wamp_registered/0

Specs

wamp_registered() :: #registered{}.

wamp_result/0

Specs

wamp_result() :: #result{}.

wamp_subscribe/0

Specs

wamp_subscribe() :: #subscribe{}.

wamp_subscribed/0

Specs

wamp_subscribed() :: #subscribed{}.

wamp_subscriber_received/0

Specs

wamp_subscriber_received() :: #subscriber_received{}.

wamp_unregister/0

Specs

wamp_unregister() :: #unregister{}.

wamp_unregistered/0

Specs

wamp_unregistered() :: #unregistered{}.

wamp_unsubscribe/0

Specs

wamp_unsubscribe() :: #unsubscribe{}.

wamp_unsubscribed/0

Specs

wamp_unsubscribed() :: #unsubscribed{}.

wamp_welcome/0

Specs

wamp_welcome() :: #welcome{}.

wamp_yield/0

Specs

wamp_yield() :: #yield{}.

Functions

close_context(Ctxt)

Specs

close_context(bondy_context:t()) -> bondy_context:t().

features()

Specs

features() -> map().

handle_message(M, Ctxt)

Specs

handle_message(M :: wamp_message(), Ctxt :: bondy_context:t()) -> ok | no_return().

Handles a wamp message. This function is called by the bondy_router module. The message might be handled synchronously (it is performed by the calling process i.e. the transport handler) or asynchronously (by sending the message to the broker worker pool).

handle_message(M, To, Opts)

Specs

handle_message(M :: wamp_publish(), To :: bondy_ref:t(), Opts :: map()) -> ok | no_return().

Handles a message sent by a peer node through the bondy_peer_wamp_relay.

is_feature_enabled(F)

Specs

is_feature_enabled(binary()) -> boolean().

Returns true if feature F is enabled by the broker.

publish(Opts, TopicUri, Args, ArgsKw, Ctxt)

Specs

publish(Opts :: map(),
        {Realm :: uri(), TopicUri :: uri()} | uri(),
        Args :: [],
        ArgsKw :: map(),
        bondy_context:t()) ->
           {ok, id()} | {error, any()}.

publish(ReqId, Opts, TopicUri, Args, ArgsKw, RealmUri)

Specs

publish(id(),
        Opts :: map(),
        {Realm :: uri(), TopicUri :: uri()} | uri(),
        Args :: list(),
        ArgsKw :: map(),
        bondy_context:t() | uri()) ->
           {ok, id()} | {error, any()}.

subscribe(RealmUri, Opts, Topic)

Specs

subscribe(RealmUri :: uri(), Opts :: map(), Topic :: uri()) ->
             {ok, id()} | {ok, id(), pid()} | {error, already_exists | any()}.

subscribe(RealmUri, Opts, Topic, SubscriberOrFun)

Specs

subscribe(RealmUri :: uri(), Opts :: map(), Topic :: uri(), SubscriberOrFun :: pid() | function()) ->
             {ok, id()} | {ok, id(), pid()} | {error, already_exists | any()}.

For internal use. If the last argument is a function, spawns a supervised instance of a bondy_subscriber by calling bondy_subscribers_sup:start_subscriber/4. The new process, calls subscribe/4 passing its pid as last argument.
If the last argument is a pid, it registers the pid as a subscriber (a.k.a a local subscription)

unsubscribe(Subscriber)

Specs

unsubscribe(pid()) -> ok | {error, not_found}.

For internal Bondy use. Terminates the process identified by Pid by bondy_subscribers_sup:terminate_subscriber/1

unsubscribe(SubsId, RealmUri)

Specs

unsubscribe(id(), bondy_context:t() | uri()) -> ok | {error, not_found}.