bondy_cidr (bondy v1.0.0-beta.16)

Types

t/0

Specs

t() :: {inet:ip_address(), non_neg_integer()}.

Functions

anchor_mask(CIDR)

Specs

anchor_mask(t()) -> t().

returns the real bottom of a netmask. Eg if 192.168.1.1/16 is provided, return 192.168.0.0/16

is_type(Term)

Specs

is_type(Term :: binary()) -> t() | no_return().

Returns true if term Term is a valid CIDR notation representation in erlang. Otherwise returns false.

mask(_)

Specs

mask(t()) -> Subnet :: binary().

match(Left, Right)

Specs

match(Left :: t(), Right :: t()) -> t() | no_return().

Returns true if Left and Right are CIDR notation representations in erlang and they match. Otherwise returns false.

parse(Bin)

Specs

parse(binary()) -> t() | no_return().

Parses a binary string representation of a CIDR notation and returns its erlang representation as a tuple t(). Fails with a badarg exception if the binary Bin is not a valid input.