bondy_context (bondy v1.0.0-beta.16)

A Bondy Context lets you access information that defines the state of an interaction. In a typical interacion, several actors or objects have a hand in what is going on e.g. bondy_session, wamp_realm, etc.
The Bondy Context is passed as an argument through the whole request-response loop to provide access to that information.

Types

encoding/0

Specs

encoding() ::
    json | msgpack | bert | erl | json_batched | msgpack_batched | bert_batched | erl_batched.

frame_type/0

Specs

frame_type() :: text | binary.

id/0

Specs

id() :: 0..9007199254740992.

maybe/1

Specs

maybe(T) :: T | undefined.

subprotocol/0

Specs

subprotocol() :: {transport(), frame_type(), encoding()}.

subprotocol_2/0

Specs

subprotocol_2() :: subprotocol() | {http, text, json | msgpack}.

t/0

Specs

t() ::
    #{id => id(),
      realm_uri => uri(),
      node => atom(),
      security_enabled => boolean(),
      session => bondy_session:t() | undefined,
      peer => bondy_session:peer(),
      authmethod => binary(),
      authid => binary(),
      is_anonymous => boolean(),
      roles => map(),
      request_id => id(),
      request_timestamp => integer(),
      request_timeout => non_neg_integer(),
      request_details => map(),
      user_info => map()}.

transport/0

Specs

transport() :: ws | raw.

uri/0

Specs

uri() :: binary().

Functions

agent(_)

Specs

agent(t()) -> binary() | undefined.

Returns the agent of the provided context or 'undefined' if there is none.

authid(_)

Specs

authid(t()) -> binary() | anonymous | undefined.

authrole(_)

Specs

authrole(t()) -> binary() | anonymous | undefined.

authroles(_)

Specs

authroles(t()) -> [binary()].

call_timeout(_)

Specs

call_timeout(t()) -> non_neg_integer().

Returns the current WAMP call timeout.

caller_details(Ctxt, Details)

Specs

caller_details(Ctxt :: t(), Details :: map()) -> map().

Returns a copy of Details` where the disclose_caller feature properties have been added from context `Ctxt.

close(Ctxt0)

Specs

close(t()) -> ok.

Closes the context. This function calls close/2 with normal as reason.

close(Ctxt, Reason)

Specs

close(t(), Reason :: normal | crash | shutdown) -> ok.

Closes the context. This function calls bondy_session:close/1 and bondy_router:close_context/1.

encoding(_)

Specs

encoding(t()) -> encoding().

Returns the encoding used by the peer of the provided context.

format_status(Opts, Ctxt)

Specs

format_status(Opts :: normal | terminate, Ctxt :: t()) -> term().

get_id(Ctxt, Scope)

Specs

get_id(Ctxt :: t(), Scope :: global | router | session) -> id().

Returns an ID based on scope. If context does not have a session the global sceop is used.

has_session(_)

Specs

has_session(t()) -> boolean().

Returns true if the context is associated with a session, false otherwise.

id(_)

Specs

id(t()) -> id().

Returns the context identifier.

is_anonymous(Ctxt)

Specs

is_anonymous(t()) -> boolean().

Returns true if the user is anonymous. In that case authid would be a random identifier assigned by Bondy.

is_feature_enabled(Ctxt, Role, Feature)

Specs

is_feature_enabled(t(), atom(), binary()) -> boolean().

Returns true if the feature Feature is enabled for role Role.

is_security_enabled(_)

Specs

is_security_enabled(t()) -> boolean().

local_context(RealmUri)

new()

Specs

new() -> t().

Initialises a new context.

new(Peer, Subprotocol)

Specs

new(bondy_session:peer(), subprotocol_2()) -> t().

node(_)

Specs

node(t()) -> atom().

Returns the peer of the provided context.

peer(_)

Specs

peer(t()) -> bondy_session:peer().

Returns the peer of the provided context.

peername(_)

Specs

peername(t()) -> binary().

Returns the peer of the provided context.

publisher_details(Ctxt, Details)

Specs

publisher_details(Ctxt :: t(), Details :: map()) -> map().

Returns a copy of Details` where the disclose_publisher feature properties have been added from context `Ctxt.

rbac_context(_)

Specs

rbac_context(t()) -> bondy_rbac:context() | undefined.

Returns the sessionId of the provided context or 'undefined' if there is none.

realm_uri(_)

Specs

realm_uri(t()) -> maybe(uri()).

Returns the realm uri of the provided context.

ref(_)

Specs

ref(t()) -> bondy_ref:t().

request_details(_)

Specs

request_details(t()) -> map().

Returns the current request details

request_id(_)

Specs

request_id(t()) -> id().

Returns the current request id.

request_timeout(_)

Specs

request_timeout(t()) -> non_neg_integer().

Returns the current request timeout.

request_timestamp(_)

Specs

request_timestamp(t()) -> integer().

Returns the current request timestamp.

reset(Ctxt)

Specs

reset(t()) -> t().

Resets the context. Returns a copy of Ctxt where the following attributes have been reset: request_id, request_timeout, request_timestamp

roles(Ctxt)

Specs

roles(t()) -> map().

Returns the roles of the provided context.

session(_)

Specs

session(t()) -> bondy_session:t() | no_return().

Fetches and returns the bondy_session for the associated sessionId.

session_id(_)

Specs

session_id(t()) -> id() | undefined.

Returns the sessionId of the provided context or 'undefined' if there is none.

set_authid(Ctxt, Val)

Specs

set_authid(t(), binary()) -> t().

set_call_timeout(Ctxt, Timeout)

Specs

set_call_timeout(t(), non_neg_integer()) -> t().

Sets the current WAMP call timeout to the provided context.

set_is_anonymous(Ctxt, Value)

Specs

set_is_anonymous(t(), boolean()) -> t().

set_peer(Ctxt, Peer)

Specs

set_peer(t(), bondy_session:peer()) -> t().

Set the peer to the provided context.

set_realm_uri(Ctxt, Uri)

Specs

set_realm_uri(t(), uri()) -> t().

Sets the realm uri of the provided context.

set_request_details(Ctxt, Details)

Specs

set_request_details(t(), map()) -> t().

Sets the current request details to the provided context.

set_request_id(Ctxt, ReqId)

Specs

set_request_id(t(), id()) -> t().

Sets the current request id to the provided context.

set_request_timeout(Ctxt, Timeout)

Specs

set_request_timeout(t(), non_neg_integer()) -> t().

Sets the current request timeout to the provided context.

set_request_timestamp(Ctxt, Timestamp)

Specs

set_request_timestamp(t(), integer()) -> t().

Sets the current request timeout to the provided context.

set_session(Ctxt, S)

Specs

set_session(t(), bondy_session:t()) -> t().

Sets the sessionId to the provided context.

set_subprotocol(Ctxt, S)

Specs

set_subprotocol(t(), subprotocol_2()) -> t().

Set the peer to the provided context.

subprotocol(_)

Specs

subprotocol(t()) -> subprotocol_2().

Returns the subprotocol of the provided context.