bondy_dealer (bondy v1.0.0-beta.16)

This module implements the capabilities of a Dealer. It is used by bondy_router.

A Dealer is one of the two roles a Router plays. In particular a Dealer is the middleman between an Caller and a Callee in an RPC interaction, i.e. it works as a generic router for remote procedure calls decoupling Callers and Callees.

Callees register the procedures they provide with Dealers. Callers initiate procedure calls first to Dealers. Dealers route calls incoming from Callers to Callees implementing the procedure called, and route call results back from Callees to Callers.

A Caller issues calls to remote procedures by providing the procedure URI and any arguments for the call. The Callee will execute the procedure using the supplied arguments to the call and return the result of the call to the Caller.

The Caller and Callee will usually implement all business logic, while the Dealer works as a generic router for remote procedure calls decoupling Callers and Callees.

Bondy does not provide message transformations to ensure stability and safety. As such, any required transformations should be handled by Callers and Callees directly (notice that a Callee can act as a middleman implementing the required transformations).

The message flow between _Callees_ and a _Dealer_ for registering and unregistering endpoints to be called over RPC involves the following messages:

REGISTER

REGISTERED

UNREGISTER

UNREGISTERED

ERROR

         ,------.          ,------.               ,------.
         |Caller|          |Dealer|               |Callee|
         `--+---'          `--+---'               `--+---'
            |                 |                      |
            |                 |                      |
            |                 |       REGISTER       |
            |                 | <---------------------
            |                 |                      |
            |                 |  REGISTERED or ERROR |
            |                 | --------------------->
            |                 |                      |
            |                 |                      |
            |                 |                      |
            |                 |                      |
            |                 |                      |
            |                 |      UNREGISTER      |
            |                 | <---------------------
            |                 |                      |
            |                 | UNREGISTERED or ERROR|
            |                 | --------------------->
         ,--+---.          ,--+---.               ,--+---.
         |Caller|          |Dealer|               |Callee|
         `------'          `------'               `------'

# Calling and Invocations

The message flow between _Callers_, a _Dealer_ and _Callees_ for calling procedures and invoking endpoints involves the following messages:

1. "CALL"

2. "RESULT"

3. "INVOCATION"

4. "YIELD"

5. "ERROR"

     sequenceDiagram
      %%{init: {'theme': 'neutral'} }%%
      Caller->>+Dealer: CALL
      Dealer->>Callee: INVOCATION
      Callee->>Dealer: YIELD | ERROR
      Dealer->>Caller: RESULT | ERROR
  

The execution of remote procedure calls is asynchronous, and there may be more than one call outstanding. A call is called outstanding (from the point of view of the _Caller_), when a (final) result or error has not yet been received by the _Caller_.

# Remote Procedure Call Ordering

Regarding *Remote Procedure Calls*, the ordering guarantees are as follows:

If _Callee A_ has registered endpoints for both *Procedure 1* and *Procedure 2*, and _Caller B_ first issues a *Call 1* to *Procedure 1* and then a *Call 2* to *Procedure 2*, and both calls are routed to _Callee A_, then _Callee A_ will first receive an invocation corresponding to *Call 1* and then *Call 2*. This also holds if *Procedure 1* and *Procedure 2* are identical.

In other words, WAMP guarantees ordering of invocations between any given _pair_ of _Caller_ and _Callee_. The current implementation relies on Distributed Erlang which guarantees message ordering betweeen processes in different nodes.

There are no guarantees on the order of call results and errors in relation to _different_ calls, since the execution of calls upon different invocations of endpoints in _Callees_ are running independently. A first call might require an expensive, long-running computation, whereas a second, subsequent call might finish immediately.

Further, if _Callee A_ registers for *Procedure 1*, the "REGISTERED" message will be sent by _Dealer_ to _Callee A_ before any "INVOCATION" message for *Procedure 1*.
There is no guarantee regarding the order of return for multiple subsequent register requests. A register request might require the _Dealer_ to do a time-consuming lookup in some database, whereas another register request second might be permissible immediately.

Types

id/0

Specs

id() :: 0..9007199254740992.

invoke_opts/0

Specs

invoke_opts() ::
    #{error_formatter := maybe(fun((Reason :: any()) -> wamp_error() | undefined)),
      call_opts := map()}.

maybe/1

Specs

maybe(T) :: T | undefined.

uri/0

Specs

uri() :: binary().

wamp_abort/0

Specs

wamp_abort() :: #abort{}.

wamp_authenticate/0

Specs

wamp_authenticate() :: #authenticate{}.

wamp_call/0

Specs

wamp_call() :: #call{}.

wamp_cancel/0

Specs

wamp_cancel() :: #cancel{}.

wamp_challenge/0

Specs

wamp_challenge() :: #challenge{}.

wamp_error/0

Specs

wamp_error() :: #error{}.

wamp_event/0

Specs

wamp_event() :: #event{}.

wamp_event_received/0

Specs

wamp_event_received() :: #event_received{}.

wamp_goodbye/0

Specs

wamp_goodbye() :: #goodbye{}.

wamp_hello/0

Specs

wamp_hello() :: #hello{}.

wamp_interrupt/0

Specs

wamp_interrupt() :: #interrupt{}.

wamp_invocation/0

Specs

wamp_invocation() :: #invocation{}.

wamp_message/0

Specs

wamp_message() ::
    wamp_hello() |
    wamp_challenge() |
    wamp_authenticate() |
    wamp_welcome() |
    wamp_abort() |
    wamp_goodbye() |
    wamp_error() |
    wamp_publish() |
    wamp_published() |
    wamp_subscribe() |
    wamp_subscribed() |
    wamp_unsubscribe() |
    wamp_unsubscribed() |
    wamp_event() |
    wamp_event_received() |
    wamp_subscriber_received() |
    wamp_call() |
    wamp_cancel() |
    wamp_result() |
    wamp_register() |
    wamp_registered() |
    wamp_unregister() |
    wamp_unregistered() |
    wamp_invocation() |
    wamp_interrupt() |
    wamp_yield().

wamp_publish/0

Specs

wamp_publish() :: #publish{}.

wamp_published/0

Specs

wamp_published() :: #published{}.

wamp_register/0

Specs

wamp_register() :: #register{}.

wamp_registered/0

Specs

wamp_registered() :: #registered{}.

wamp_result/0

Specs

wamp_result() :: #result{}.

wamp_subscribe/0

Specs

wamp_subscribe() :: #subscribe{}.

wamp_subscribed/0

Specs

wamp_subscribed() :: #subscribed{}.

wamp_subscriber_received/0

Specs

wamp_subscriber_received() :: #subscriber_received{}.

wamp_unregister/0

Specs

wamp_unregister() :: #unregister{}.

wamp_unregistered/0

Specs

wamp_unregistered() :: #unregistered{}.

wamp_unsubscribe/0

Specs

wamp_unsubscribe() :: #unsubscribe{}.

wamp_unsubscribed/0

Specs

wamp_unsubscribed() :: #unsubscribed{}.

wamp_welcome/0

Specs

wamp_welcome() :: #welcome{}.

wamp_yield/0

Specs

wamp_yield() :: #yield{}.

Functions

callees(RealmUri)

Specs

callees(RealmUri :: uri()) -> [map()] | no_return().

callees(RealmUri, ProcedureUri)

Specs

callees(RealmUri :: uri(), ProcedureUri :: uri()) -> [map()] | no_return().

callees(RealmUri, ProcedureUri, Opts)

Specs

callees(RealmUri :: uri(), ProcedureUri :: uri(), Opts :: map()) -> [map()] | no_return().

close_context(Ctxt)

Specs

close_context(bondy_context:t()) -> bondy_context:t().

features()

Specs

features() -> map().

handle_message(M, Ctxt)

Specs

handle_message(M :: wamp_message(), Ctxt :: map()) -> ok | no_return().

handle_message(Invocation, To, Opts)

Specs

handle_message(wamp_message(), To :: bondy_ref:t(), Opts :: map()) -> ok | no_return().

Handles inbound messages received from a relay i.e. a cluster peer node or bridge_relay i.e. edge client or server.

is_feature_enabled(F)

Specs

is_feature_enabled(binary()) -> boolean().

register(Procedure, Opts, Ref)

Specs

register(Procedure :: uri(), Opts :: map(), Ref :: bondy_ref:t() | bondy_context:t()) ->
            {ok, id()} | {ok, id(), pid()} | {error, already_exists | any()} | no_return().

Creates a local registration. If the registration is done using a callback module, only the invoke single strategy can be used (i.e. shared_registration and sharded_registration are also disabled). Also the callback module needs to conform to the wamp_api_callback behaviour, otherwise the call fails with a badarg exception.

unregister(Callee)

Specs

unregister(pid()) -> ok | {error, not_found}.

For internal Bondy use. Terminates the process identified by Pid by bondy_subscribers_sup:terminate_subscriber/1

unregister(RegId, RealmUri)

Specs

unregister(RegId :: id(), bondy_context:t() | uri()) -> ok | {error, not_found}.