bondy_edge_uplink_client (bondy v1.0.0-beta.16)

EARLY DRAFT implementation of the client-side connection between and edge node (client) and a remote/core node (server).

Types

maybe/1

Specs

maybe(T) :: T | undefined.

sessions/0

Specs

sessions() :: #{bondy_session:id() => bondy_edge_session:t()}.

t/0

Specs

t() :: #state{}.

uri/0

Specs

uri() :: binary().

Functions

callback_mode()

code_change(OldVsn, StateName, StateData, Extra)

connected(EventType, Msg, State)

The edge router established the uplink connection with the core router. At this point the edge router has noy yet joined any realms.

connecting(EventType, Msg, State)

The edge router is trying to establish an uplink connection to the core router. If reconnect is configured it will retry using the reconnect options defined.

forward(Pid, Msg, SessionId)

Specs

forward(Pid :: pid(), Msg :: any(), SessionId :: bondy_session:id()) -> ok.

init(_)

start_link(Transport, Endpoint, Opts)

terminate(Reason, StateName, State)

Specs

terminate(term(), atom(), t()) -> term().