bondy_event_manager (bondy v1.0.0-beta.16)

This module is used by Bondy to manage event handlers and notify them of events. It implements the "watched" handler capability i.e. add_watched_handler/2,3, swap_watched_handler/2,3. In addition, this module mirrors most of the gen_event API and adds variants with two arguments were the first argument is the defaul event manager (bondy_event_manager).

       +---------------------------------------+
       |                                       |
       |          bondy_event_manager          |
       |                                       |
       +---------------------------------------+
                           |
                           |
                           v
       +---------------------------------------+
       |                                       |
       |    bondy_event_handler_watcher_sup    |
       |                                       |
       +---------------------------------------+
                           |
                           +--------------------------------+
                           |                                |
        +---------------------------------------+       +---+---+
        |                                       |       |       |
        |     bondy_event_handler_watcher 1     |       |   N   |
        |                                       |       |       |
        +---------------------------------------+       +-------+
 
                        simple_one_for_one

Functions

add_handler(Handler, Args)

Adds an event handler. Calls gen_event:add_handler(?MODULE, Handler, Args).

add_handler(Manager, Handler, Args)

Adds an event handler. Calls gen_event:add_handler(Manager, Handler, Args).

add_sup_handler(Handler, Args)

Adds a supervised event handler, but also supervises the connection between the event handler and the calling process. Calls gen_event:add_sup_handler(?MODULE, Handler, Args). %% Use this call if you want the event manager to remove the handler when the calling process terminates.

add_sup_handler(Manager, Handler, Args)

Adds a supervised event handler, but also supervises the connection between the event handler and the calling process. Calls gen_event:add_sup_handler(?MODULE, Handler, Args). Use this call if you want the event manager to remove the handler when the calling process terminates.

add_watched_handler(Handler, Args)

Adds a watched event handler. As opposed to add_sup_handler/2 which supervises the calling process, this function calls bondy_event_handler_watcher_sup:start_watcher(Handler, Args) which spawns a supervised process (bondy_event_handler_watcher) which calls add_sup_handler/2. If the handler crashes, bondy_event_handler_watcher will re-install it in the event manager.

add_watched_handler(Manager, Handler, Args)

Adds a supervised event handler. As opposed to add_sup_handler/2 which monitors the calling process, this function calls bondy_event_handler_watcher_sup:start_watcher(Handler, Args) which spawns a supervised process (bondy_event_handler_watcher) which calls add_sup_handler/2. If the handler crashes, bondy_event_handler_watcher will re-install it in the event manager.

notify(Event)

A util function. Equivalent to calling notify(bondy_event_manager, Event)

notify(Manager, Event)

A util function. Equivalent to calling gen_event:notify(bondy_event_manager, Event)

swap_handler(OldHandler, NewHandler)

A util function. Equivalent to calling swap_handler(bondy_event_manager, OldHandler, NewHandler)

swap_handler(Manager, OldHandler, NewHandler)

A util function. Equivalent to calling gen_event:swap_handler/3

swap_sup_handler(OldHandler, NewHandler)

A util function. Equivalent to calling swap_sup_handler(bondy_event_manager, OldHandler, NewHandler)

swap_sup_handler(Manager, OldHandler, NewHandler)

A util function. Equivalent to calling gen_event:swap_sup_handler/3

swap_watched_handler(OldHandler, NewHandler)

A util function. Equivalent to calling swap_watched_handler(bondy_event_manager, OldHandler, NewHandler)

swap_watched_handler(Manager, OldHandler, NewHandler)

Replaces an event handler in event manager Manager in the same way as swap_sup_handler/3. However, this function calls bondy_event_handler_watcher_sup:start_watcher(Handler, Args) which spawns a supervised process (bondy_event_handler_watcher) which is the one calling calls swap_sup_handler/2. If the handler crashes or terminates with a reason other than normal or shutdown, bondy_event_handler_watcher will re-install it in the event manager.

sync_notify(Event)

A util function. Equivalent to calling sync_notify(bondy_event_manager, Event)

sync_notify(Manager, Event)

A util function. Equivalent to calling gen_event:sync_notify(bondy_event_manager, Event)