bondy_http_gateway (bondy v1.0.0-beta.16)

Types

listener/0

Specs

listener() :: api_gateway_http | api_gateway_https | admin_api_http | admin_api_https.

Functions

apply_config()

Specs

apply_config() -> ok | {error, invalid_specification_format | any()}.

Parses the apis provided by the configuration file ('bondy.api_gateway.config_file'), stores the apis in the metadata store. This call does not load the spec in the HTTP server dispatch tables.

code_change(OldVsn, State, Extra)

delete(Id)

Specs

delete(binary()) -> ok.

Deletes the API Specification object identified by Id. Notice: Triggers a rebuild of the Cowboy dispatch tables.

dispatch_table(Listener)

Specs

dispatch_table(listener()) -> any().

Returns the current dispatch configured in Cowboy for the given Listener.

handle_call(Event, From, State)

handle_cast(Event, State)

handle_info(Info, State)

init(_)

list()

Specs

list() -> [ParsedSpec :: map()].

Returns the list of all stored API Specification objects.

load(Term)

Specs

load(file:filename() | map()) -> ok | {error, invalid_specification_format | any()}.

Parses the provided Spec, adds it to the store and calls rebuild_dispatch_tables/0.

lookup(Id)

Specs

lookup(binary()) -> map() | {error, not_found}.

Returns the API Specification object identified by Id.

rebuild_dispatch_tables()

Loads all configured API specs from the metadata store and rebuilds the Cowboy dispatch table by calling cowboy_router:compile/1 and updating the environment.

resume_admin_listeners()

Specs

resume_admin_listeners() -> ok.

resume_listeners()

Specs

resume_listeners() -> ok.

start_admin_listeners()

start_link()

start_listeners()

Specs

start_listeners() -> ok.

Conditionally start the public http and https listeners based on the configuration. This will load any default Bondy api specs. Notice this is not the way to start the admin api listeners see start_admin_listeners() for that.

stop_admin_listeners()

Specs

stop_admin_listeners() -> ok.

stop_listeners()

Specs

stop_listeners() -> ok.

suspend_admin_listeners()

Specs

suspend_admin_listeners() -> ok.

suspend_listeners()

Specs

suspend_listeners() -> ok.

terminate(Reason, State)