bondy_http_gateway_api_spec_parser (bondy v1.0.0-beta.16)

This module implements a parser for the Bondy API Specification Format which is used to dynamically define one or more RESTful APIs.

The parser uses maps_utils:validate and a number of validation specifications and as a result is its error reporting is not great but does its job. The plan is to replace this with an ad-hoc parser to give the user more useful error information.

## API Specification Format

### Host Spec

### API Spec

### Version Spec

### Path Spec

### Method Spec

### Action Spec
### Action Spec

Types

route_match/0

Specs

route_match() :: '_' | iodata().

route_path/0

Specs

route_path() :: {Path :: route_match(), Handler :: module(), Opts :: any()}.

route_rule/0

Specs

route_rule() :: {Host :: route_match(), Paths :: [route_path()]}.

scheme_rule/0

Specs

scheme_rule() ::
    {Scheme :: binary(),
     Host :: route_match(),
     Realm :: binary(),
     Path :: route_match(),
     Handler :: module(),
     Opts :: any()}.

Cowboy types

Functions

dispatch_table(API)

Specs

dispatch_table([map()] | map()) -> [{Scheme :: binary(), [route_rule()]}] | no_return().

Given a valid API Spec or list of Specs returned by parse/1, dynamically generates a cowboy dispatch table.

dispatch_table(API, RulesToAdd)

Specs

dispatch_table([map()] | map(), [route_rule()]) ->
                  [{Scheme :: binary(), [route_rule()]}] | no_return().

Given a list of valid API Specs returned by parse/1 and dynamically generates the cowboy dispatch table. Notice this does not update the cowboy dispatch table. You will need to do that yourself.

from_file(Filename)

Specs

from_file(file:filename()) -> {ok, any()} | {error, any()}.

Loads a file and calls parse/1.

parse(Map)

Specs

parse(map()) -> map() | {error, invalid_specification_format} | no_return().

Parses the Spec map returning a new valid spec where all defaults have been applied and all variables have been replaced by either a value of a promise. Fails with in case the Spec is invalid.
Variable replacepement is performed using our "mop" library.