bondy_http_gateway_rest_handler (bondy v1.0.0-beta.16)

This module implements a generic Cowboy rest handler that handles a resource specified using the Bondy API Gateway Specification Format (BAGS), a JSON-based format for describing, producing and consuming RESTful Web Services using Bondy.

For every path defined in a BAGS file, Bondy will configure and install a Cowboy route using this module. The initial state of the module responds to a contract between this module and the bondy_http_gateway_api_spec_parser and contains the parsed and preprocessed definition of the paths specification which this module uses to dynamically implement its behaviour.
See bondy_http_gateway for a detail description of the Bondy API Gateway Specification Format.

Types

state/0

Specs

state() ::
    #{api_spec => map(),
      api_context => map(),
      body_evaluated => boolean(),
      session => any(),
      realm_uri => binary(),
      deprecated => boolean(),
      security => map(),
      is_anonymous => boolean(),
      authid => binary() | undefined,
      encoding => binary() | json | msgpack}.

Functions

accept(Req0, St0)

allowed_methods(Req, St)

content_types_accepted(Req, St)

content_types_provided(Req, St)

delete_completed(Req, St)

delete_resource(Req0, St0)

from_form_urlencoded(Req, St)

from_json(Req, St)

from_msgpack(Req, St)

init(Req, St0)

is_authorized(Req0, St0)

languages_provided(Req, St)

options(Req, St)

previously_existed(Req, St)

rate_limited(Req, St)

resource_exists(Req, St)

to_json(Req, St)

to_msgpack(Req, St)