bondy_oauth2 (bondy v1.0.0-beta.16)

The following table documents the storage layout in plum_db of the token data and its indices:
|Datum|Prefix|Key|Value| |---|---|---|---| |Refresh Token|{oauth2_refresh_tokens, Realm ++ "," ++ Issuer}|Token| TokenData| |Refresh Token Index|{oauth2_refresh_tokens, Realm ++ "," ++ Issuer}|Token|Issuer| |Refresh Token Index|{oauth2_refresh_tokens, Realm ++ "," ++ Issuer ++ "," ++ Username}|DeviceId| Token|

Types

error/0

Specs

error() :: oauth2_invalid_grant | no_such_realm.

grant_type/0

Specs

grant_type() :: client_credentials | password | authorization_code.

token_data/0

Specs

token_data() :: #bondy_oauth2_token{}.

token_type/0

Specs

token_type() :: access_token | refresh_token.

Functions

decode_jwt(JWT)

Specs

decode_jwt(binary()) -> map().

issue_token(GrantType, RealmUri, Issuer, Username, Groups, Meta)

Specs

issue_token(token_type(), bondy_realm:uri(), binary(), binary(), [binary()], map()) ->
               {ok, AccessToken :: binary(), RefreshToken :: binary(), Claims :: map()} |
               {error, any()}.

Generates an access token and a refresh token. The access token is a JWT whereas the refresh token is a binary.
The function stores the refresh token in the store and creates a number of store indices.

issued_at(Bondy_oauth2_token)

Specs

issued_at(token_data()) -> pos_integer().

Returns the timestamp for the token data.

issuer(TokenData)

Specs

issuer(TokenData :: token_data()) -> binary().

Returns the issuer a.k.a. ClientId of the token data.

lookup_token(Realm, Issuer, Token)

Specs

lookup_token(Realm :: bondy_realm:uri(), Issuer :: binary(), Token :: binary()) ->
                token_data() | {error, not_found}.

Returns the data token_data() associated with Token or the tuple {error, not_found}.

rebuild_token_indices(RealmUri, Issuer0)

Rebuilds refresh_token indices. This function is used for db maintenance.

refresh_token(Realm, Issuer, Token)

Specs

refresh_token(Realm :: bondy_realm:uri(), Issuer :: binary(), Token :: binary()) ->
                 {ok, AccessToken :: binary(), RefreshToken :: binary(), Claims :: map()} |
                 {error, oauth2_invalid_grant}.

After refreshing a token, the previous refresh token will be revoked

revoke_dangling_tokens(RealmUri, Issuer0)

Removes all refresh tokens whose user has been removed. This function is used for db maintenance.

revoke_refresh_token(Realm, IssuerOrData, Token)

Specs

revoke_refresh_token(Realm :: bondy_realm:uri(),
                     IssuerOrData :: binary() | token_data(),
                     Token :: binary()) ->
                        ok.

Removes a refresh token from store using an index to match the function arguments. This also removes all store indices.

revoke_refresh_token(RealmUri, Issuer, Username, DeviceId)

Specs

revoke_refresh_token(bondy_realm:uri(),
                     Issuer :: binary(),
                     Username :: binary(),
                     DeviceId :: binary()) ->
                        ok.

Removes a refresh token from store using an index to match the function arguments. This also removes all store indices.

revoke_refresh_tokens(RealmUri, Username)

Specs

revoke_refresh_tokens(bondy_realm:uri(), Username :: binary()) -> ok.

revoke_refresh_tokens(RealmUri, Issuer, Username)

Specs

revoke_refresh_tokens(bondy_realm:uri(), Issuer :: binary(), Username :: binary()) -> ok.

revoke_token(Hint, RealmUri, Issuer, TokenOrUsername)

Specs

revoke_token(Hint :: token_type() | undefined,
             bondy_realm:uri(),
             Issuer :: binary(),
             TokenOrUsername :: binary()) ->
                ok | {error, unsupported_operation}.

revoke_token(Hint, RealmUri, Issuer, Username, DeviceId)

Specs

revoke_token(Hint :: token_type() | undefined,
             bondy_realm:uri(),
             Issuer :: binary(),
             Username :: binary(),
             DeviceId :: non_neg_integer()) ->
                ok | {error, unsupported_operation}.

revoke_tokens(Hint, Realm, Username)

Specs

revoke_tokens(Hint :: token_type() | undefined, Realm :: bondy_realm:uri(), Username :: binary()) ->
                 ok | {error, unsupported_operation}.

revoke_tokens(Hint, RealmUri, Issuer, Username)

Specs

revoke_tokens(Hint :: token_type() | undefined,
              bondy_realm:uri(),
              Issuer :: binary(),
              Username :: binary()) ->
                 ok | {error, unsupported_operation}.

verify_jwt(RealmUri, JWT)

Specs

verify_jwt(binary(), binary()) -> {ok, map()} | {error, error()}.

verify_jwt(RealmUri, JWTString, MatchSpec)

Specs

verify_jwt(RealmUri :: binary(), JWTString :: binary(), MatchSpec :: map()) ->
              {ok, map()} | {error, error()}.