bondy_password (bondy v1.0.0-beta.16)

A Password object stores a fixed size salted hash of a user's password and all the metadata required to re-compute the salted hash for comparing a user input and for implementing several password-based authentication protocols.
At the moment this module supports two protocols: * WAMP Challenge-Response Authentication (CRA), and * Salted Challenge-Response Authentication Mechanism (SCRAM)

Types

data/0

Specs

data() :: bondy_password_cra:data() | bondy_password_scram:data().

future/0

Specs

future() :: fun((opts()) -> t()).

opts/0

Specs

opts() :: #{protocol := protocol(), params := params()}.

params/0

Specs

params() :: bondy_password_cra:params() | bondy_password_scram:params().

protocol/0

Specs

protocol() :: cra | scram.

t/0

Specs

t() ::
    #{type := password,
      version := binary(),
      protocol := protocol(),
      params := params(),
      data := data()}.

Functions

data(_)

Specs

data(t()) -> data().

default_opts()

Specs

default_opts() -> opts().

default_opts(Protocol)

Specs

default_opts(protocol()) -> opts().

from_term(Term)

Specs

from_term(Term :: proplist:proplist() | map()) -> t().

future(Password)

Specs

future(binary()) -> future().

Creates a functional object that takes a single argument Opts :: opts() that when applied calls new(Password, Opts).

This is used for two reasons: 1. to encapsulate the string value of the password avoiding exposure i.e. via logs; and 2. To delay the processing of the password until the value for Opts is known.

Password must be a binary with a minimum size of 6 bytes and a maximum size of 256 bytes, otherwise fails with error invalid_password.

Example:

     erlang
  > F = bondy_password:future(<<"MyBestKeptSecret">>).
  > bondy_password:new(F, Opts).

hash_length(PW)

Specs

hash_length(t()) -> pos_integer().

is_type(_)

Specs

is_type(t()) -> boolean().

new(Future, Opts)

Specs

new(binary() | future(), opts()) -> t() | no_return().

Hash a plaintext password Password and the protocol and protocol params defined in options Opts, returning t().
Password must be a binary with a minimum size of 6 bytes and a maximum size of 256 bytes, otherwise fails with error invalid_password.

opts_validator()

Specs

opts_validator() -> map().

params(_)

Specs

params(t()) -> params().

protocol(_)

Specs

protocol(t()) -> protocol() | undefined.

replace(Password, PW)

Specs

replace(Password :: binary() | future(), PW :: t()) -> t() | no_return().

Returns a new password object from String applying the same protocol and params found in password PWD.

upgrade(String, T0)

Specs

upgrade(String :: tuple() | binary(), T0 :: map() | proplists:proplist()) ->
           {true, T1 :: t()} | false.

verify_hash(Hash, Password)

Specs

verify_hash(Hash :: binary(), Password :: t()) -> boolean().

verify_string(String, Password)

Specs

verify_string(String :: binary(), Password :: t()) -> boolean().