bondy_password_scram (bondy v1.0.0-beta.16)

This module provides the functions and algorithms to operate with the Salted Challenge-Reponse Mechanism data structures.

Types

data/0

Specs

data() :: #{salt := binary(), stored_key := binary(), server_key := binary()}.

hash_fun/0

Specs

hash_fun() :: sha256.

kdf/0

Specs

kdf() :: pbkdf2 | argon2id13.

params/0

Specs

params() ::
    #{kdf := kdf(),
      iterations := non_neg_integer(),
      memory => non_neg_integer(),
      hash_function := hash_fun(),
      hash_length := non_neg_integer(),
      salt_length := non_neg_integer()}.

Functions

auth_message(AuthId, ClientNonce, ServerNonce, Salt, Iterations)

auth_message(AuthId, ClientNonce, ServerNonce, Salt, Iterations, CBindName, CBindData)

check_proof(ProvidedProof, ClientProof, ClientSignature, StoredKey)

Specs

check_proof(ProvidedProof :: binary(),
            ClientProof :: binary(),
            ClientSignature :: binary(),
            StoredKey :: binary()) ->
               boolean().

client_key(SaltedPassword)

Specs

client_key(SaltedPassword :: binary()) -> ClientKey :: binary().

client_proof(Key, Signature)

Specs

client_proof(Key :: binary(), Signature :: binary()) -> ClientProof :: binary().

computes the client proof out of the client key Key and the client signature Signature. See client_key/2 and client_signature/2 respectively.

client_signature(StoredKey, AuthMessage)

Specs

client_signature(StoredKey :: binary(), AuthMessage :: binary()) -> ClientSignature :: binary().

hash_function()

Specs

hash_function() -> atom().

hash_length()

Specs

hash_length() -> integer().

new(String, Params0, Builder)

Specs

new(binary(), params(), fun((data(), params()) -> bondy_password:t())) ->
       bondy_password:t() | no_return().

recovered_client_key(ClientProof, ClientSignature)

Specs

recovered_client_key(ClientProof :: binary(), ClientSignature :: binary()) ->
                        RecoveredClientKey :: binary().

recovered_stored_key(RecoveredClientKey)

Specs

recovered_stored_key(RecoveredClientKey :: binary()) -> RecoveredStoredKey :: binary().

salt()

Specs

salt() -> Salt :: binary().

salt_length()

Specs

salt_length() -> integer().

salted_password(Password, Salt, Params)

Specs

salted_password(Password :: binary(), Salt :: binary(), Params :: params()) ->
                   SaltedPassword :: binary().

server_key(SaltedPassword)

Specs

server_key(SaltedPassword :: binary()) -> ServerKey :: binary().

server_nonce(ClientNonce)

Specs

server_nonce(ClientNonce :: binary()) -> ServerNonce :: binary().

server_signature(ServerKey, AuthMessage)

Specs

server_signature(ServerKey :: binary(), AuthMessage :: binary()) -> ClientSignature :: binary().

stored_key(ClientKey)

Specs

stored_key(ClientKey :: binary()) -> StoredKey :: binary().

validate_params(Params)

Specs

validate_params(Params :: params()) -> Validated :: params() | no_return().

verify_string(String, Data, Params)

Specs

verify_string(binary(), data(), params()) -> boolean().