bondy_peer_discovery_agent behaviour (bondy v1.0.0-beta.16)

This state machine is reponsible for discovering Bondy cluster peers using the defined implementation backend (callback module).

Its behaviour can be configured using the cluster.peer_discovery family of bondy.conf options.
If the agent is enabled (cluster.peer_discovery.enabled) and the Bondy node has not yet joined a cluster, it will lookup for peers using the defined implementation backend callback module (cluster.peer_discovery.type).

Callbacks

init/1

Specs

init(Opts :: map()) -> {ok, State :: any()} | {error, Reason :: any()}.

lookup/2

Specs

lookup(State :: any(), Timeout :: timeout()) ->
          {ok, [bondy_peer_service:peer()], NewState :: any()} |
          {error, Reason :: any(), NewState :: any()}.

Functions

callback_mode()

code_change(OldVsn, StateName, State, Extra)

disable()

Specs

disable() -> boolean().

disabled(EventType, EventContent, State)

discovering(EventType, EventContent, State)

In this state the agent uses the callback module to discover peers by calling its lookup/2 callback.

enable()

Specs

enable() -> boolean().

format_status(Opts, _)

init(_)

joined(EventType, EventContent, State)

lookup()

start()

Specs

start() -> {ok, pid()} | ignore | {error, term()}.

start_link()

Specs

start_link() -> {ok, pid()} | ignore | {error, term()}.

terminate(Reason, StateName, State)