bondy_peer_discovery_dns_agent (bondy v1.0.0-beta.16)

An implementation of the bondy_peer_discovery_agent behaviour that uses DNS for service discovery.

It is enabled by using the following options in the bondy.conf file

     bash
  cluster.peer_discovery_agent.type = bondy_peer_discovery_dns_agent
  cluster.peer_discovery_agent.config.service_name = my-service-name
Where service_name is the service to be used by the DNS lookup.

Functions

init(Opts)

Specs

init(Opts :: map()) -> {ok, State :: any()} | {error, Reason :: any()}.

lookup(State, Timeout)

Specs

lookup(State :: any(), timeout()) ->
          {ok, [bondy_peer_service:peer()], NewState :: any()} |
          {error, Reason :: any(), NewState :: any()}.