bondy_peer_wamp_relay (bondy v1.0.0-beta.16)

A gen_server that forwards INVOCATION (their RESULT or ERROR), INTERRUPT and EVENT messages between WAMP clients connected to different Bondy peers (nodes).

  +-------------------------+                    +-------------------------+
  |         node_1          |                    |         node_2          |
  |                         |                    |                         |
  |                         |                    |                         |
  | +---------------------+ |    cast_message    | +---------------------+ |
  | |partisan_peer_service| |                    | |partisan_peer_service| |
  | |      _manager       |<+--------------------+>|      _manager       | |
  | |                     | |                    | |                     | |
  | +---------------------+ |                    | +---------------------+ |
  |    ^          |         |                    |         |          ^    |
  |    |          v         |                    |         v          |    |
  |    |  +---------------+ |                    | +---------------+  |    |
  |    |  |bondy_peer_wamp| |                    | |bondy_peer_wamp|  |    |
  |    |  |    _relay     | |                    | |    _relay     |  |    |
  |    |  |               | |                    | |               |  |    |
  |    |  +---------------+ |                    | +---------------+  |    |
  |    |          |         |                    |         |          |    |
  |    |          |         |                    |         |          |    |
  |    |          |         |                    |         |          |    |
  |    |          v         |                    |         v          |    |
  | +---------------------+ |                    | +---------------------+ |
  | |       Worker        | |                    | |       Worker        | |
  | |    (router_pool)    | |                    | |    (router_pool)    | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | +---------------------+ |                    | +---------------------+ |
  |         ^    |          |                    |          |   ^          |
  |         |    |          |                    |          |   |          |
  |         |    v          |                    |          v   |          |
  | +---------------------+ |                    | +---------------------+ |
  | |bondy_wamp_*_handler | |                    | |bondy_wamp_*_handler | |
  | |                     | |                    | |                     | |
  | |                     | |                    | |                     | |
  | +---------------------+ |                    | +---------------------+ |
  |         ^    |          |                    |          |   ^          |
  |         |    |          |                    |          |   |          |
  +---------+----+----------+                    +----------+---+----------+
            |    |                                          |   |
            |    |                                          |   |
       CALL |    | RESULT | ERROR                INVOCATION |   | YIELD
            |    |                                          |   |
            |    v                                          v   |
  +-------------------------+                    +-------------------------+
  |         Caller          |                    |         Callee          |
  |                         |                    |                         |
  |                         |                    |                         |
  +-------------------------+                    +-------------------------+

Types

id/0

Specs

id() :: 0..9007199254740992.

uri/0

Specs

uri() :: binary().

wamp_abort/0

Specs

wamp_abort() :: #abort{}.

wamp_authenticate/0

Specs

wamp_authenticate() :: #authenticate{}.

wamp_call/0

Specs

wamp_call() :: #call{}.

wamp_cancel/0

Specs

wamp_cancel() :: #cancel{}.

wamp_challenge/0

Specs

wamp_challenge() :: #challenge{}.

wamp_error/0

Specs

wamp_error() :: #error{}.

wamp_event/0

Specs

wamp_event() :: #event{}.

wamp_event_received/0

Specs

wamp_event_received() :: #event_received{}.

wamp_goodbye/0

Specs

wamp_goodbye() :: #goodbye{}.

wamp_hello/0

Specs

wamp_hello() :: #hello{}.

wamp_interrupt/0

Specs

wamp_interrupt() :: #interrupt{}.

wamp_invocation/0

Specs

wamp_invocation() :: #invocation{}.

wamp_message/0

Specs

wamp_message() ::
    wamp_hello() |
    wamp_challenge() |
    wamp_authenticate() |
    wamp_welcome() |
    wamp_abort() |
    wamp_goodbye() |
    wamp_error() |
    wamp_publish() |
    wamp_published() |
    wamp_subscribe() |
    wamp_subscribed() |
    wamp_unsubscribe() |
    wamp_unsubscribed() |
    wamp_event() |
    wamp_event_received() |
    wamp_subscriber_received() |
    wamp_call() |
    wamp_cancel() |
    wamp_result() |
    wamp_register() |
    wamp_registered() |
    wamp_unregister() |
    wamp_unregistered() |
    wamp_invocation() |
    wamp_interrupt() |
    wamp_yield().

wamp_publish/0

Specs

wamp_publish() :: #publish{}.

wamp_published/0

Specs

wamp_published() :: #published{}.

wamp_register/0

Specs

wamp_register() :: #register{}.

wamp_registered/0

Specs

wamp_registered() :: #registered{}.

wamp_result/0

Specs

wamp_result() :: #result{}.

wamp_subscribe/0

Specs

wamp_subscribe() :: #subscribe{}.

wamp_subscribed/0

Specs

wamp_subscribed() :: #subscribed{}.

wamp_subscriber_received/0

Specs

wamp_subscriber_received() :: #subscriber_received{}.

wamp_unregister/0

Specs

wamp_unregister() :: #unregister{}.

wamp_unregistered/0

Specs

wamp_unregistered() :: #unregistered{}.

wamp_unsubscribe/0

Specs

wamp_unsubscribe() :: #unsubscribe{}.

wamp_unsubscribed/0

Specs

wamp_unsubscribed() :: #unsubscribed{}.

wamp_welcome/0

Specs

wamp_welcome() :: #welcome{}.

wamp_yield/0

Specs

wamp_yield() :: #yield{}.

Functions

async_forward(Ref, Msg, Opts)

Specs

async_forward(Ref :: bondy_ref:t(), Msg :: wamp_message(), Opts :: map()) ->
                 {ok, id()} | no_return().

broadcast(From, Nodes, M, Opts)

Specs

broadcast(bondy_ref:t(), [node()], wamp_message(), map()) ->
             {ok, Good :: [node()], Bad :: [node()]}.

code_change(OldVsn, State, Extra)

forward(Node, Msg)

forward(Node, Msg, Opts)

Specs

forward(Node :: node, Msg :: wamp_message(), Opts :: map()) -> ok.

Forwards a wamp message to a peer (cluster node). It returns ok when the remote bondy_peer_wamp_relay acknoledges the reception of the message, but it does not imply the message handler has actually received the message.

This only works for PUBLISH, ERROR, INTERRUPT, INVOCATION and RESULT WAMP message types. It will fail with an exception if another type is passed as the third argument.
This is equivalent to calling async_forward/3 and then yield/2.

handle_call(Event, From, State)

handle_cast(Event, State)

handle_info(Info, State)

init(_)

receive_ack(Id, Timeout)

Specs

receive_ack(bondy_ref:t(), timeout()) -> ok | no_return().

ref(RealmUri)

Specs

ref(RealmUri :: uri()) -> bondy_ref:t().

start_link()

Specs

start_link() -> {ok, pid()} | ignore | {error, term()}.

terminate(Reason, State)