bondy_rbac (bondy v1.0.0-beta.16)

### WAMP Permissions:

* "wamp.register" * "wamp.unregister" * "wamp.call" * "wamp.cancel" * "wamp.subscribe" * "wamp.unsubscribe" * "wamp.publish" * "wamp.disclose_caller" * "wamp.disclose_publisher"

### Reserved Names Reserved names are role (user or group) or resource names that act as keywords in RBAC in either binary or atom forms and thus cannot be used.

The following is the list of all reserved names.
* all - group * anonymous - the anonymous user and group * any - use to denote a resource * from - use to denote a resource * on - not used * to - use to denote a resource %% **Note:** Usernames and group names are stored in lower case. All functions in this module are case sensitice so when using the functions in this module make sure the inputs you provide are in lowercase to. If you need to convert your input to lowercase use string:casefold/1.

Types

context/0

Specs

context() :: #bondy_rbac_context{}.

grant/0

Specs

grant() :: {normalised_resource(), [Permission :: permission()]}.

normalised_resource/0

Specs

normalised_resource() :: any | {Uri :: uri(), MatchStrategy :: binary()}.

permission/0

Specs

permission() :: binary().

request/0

Specs

request() ::
    #{type := request,
      roles := [rolename()],
      permissions := [binary()],
      resources := [normalised_resource()]}.

request_data/0

Specs

request_data() :: map().

resource/0

Specs

resource() :: any | binary() | #{uri := binary(), strategy := binary()} | normalised_resource().

rolename/0

Specs

rolename() :: all | bondy_rbac_user:username() | bondy_rbac_group:name().

uri/0

Specs

uri() :: binary().

Functions

authorize(Permission, Ctxt)

Specs

authorize(binary(), bondy_context:t() | context()) -> ok | no_return().

Returns 'ok' or an exception.

authorize(Permission, Resource, Bondy_rbac_context)

Specs

authorize(binary(), binary() | any, bondy_context:t() | context()) -> ok | no_return().

Returns 'ok' or an exception.

get_anonymous_context(Ctxt)

Specs

get_anonymous_context(Ctxt :: bondy_context:t()) -> context().

get_anonymous_context(RealmUri, Username)

get_context(Ctxt)

Specs

get_context(Ctxt :: bondy_context:t()) -> context().

get_context(RealmUri, Username)

Contexts are only valid until the GRANT epoch changes, and it will change whenever a GRANT or a REVOKE is performed. This is a little coarse grained right now, but it'll do for the moment.

grant(RealmUri, Request)

Specs

grant(RealmUri :: uri(), Request :: request() | map()) ->
         ok | {error, Reason :: any()} | no_return().

**Use cases**

  grant <permissions> on any to all|{<user>|<group>[,...]}
  grant <permissions> on {<resource>, <exact|prefix|wildcard>} to all|{<user>|<group>[,...]}

grants(RealmUri, Opts)

Specs

grants(RealmUri :: uri(), Opts :: map()) -> [grant()].

Returns the local grants assigned in realm RealmUri. This function does not use protypical inheritance.

grants(RealmUri, Name, RoleType)

Specs

grants(RealmUri :: uri(), Name :: binary(), RoleType :: user | group) -> [grant()].

group_grants(RealmUri, Name)

Specs

group_grants(RealmUri :: uri(), Name :: binary()) -> [grant()].

is_reserved_name(Term)

Specs

is_reserved_name(Term :: binary() | atom()) -> boolean() | no_return().

Returns true if term is a reserved name in binary or atom form.

**Reserved names:**
* all * anonymous * any * from * on * to

normalize_name(Term)

Specs

normalize_name(Term :: binary() | atom()) -> boolean() | no_return().

Normalizes the utf8 binary Bin into a Normalized Form of compatibly equivalent Decomposed characters according to the Unicode standard and converts it to a case-agnostic comparable string.

refresh_context(Ctxt)

Specs

refresh_context(Ctxt :: bondy_context:t()) -> {boolean(), context()}.

request(Data)

Specs

request(Data :: request_data()) -> Request :: request() | no_return().

Validates the data for a grant or revoke request.

revoke(RealmUri, Request)

Specs

revoke(RealmUri :: uri(), Request :: request() | map()) ->
          ok | {error, Reason :: any()} | no_return().

revoke_group(RealmUri, Name)

revoke_user(RealmUri, Username)

user_grants(RealmUri, Username)

Specs

user_grants(RealmUri :: uri(), Username :: binary()) -> [grant()].