bondy_rbac_group (bondy v1.0.0-beta.16)

**Note:** Usernames and group names are stored in lower case. All functions in this module are case sensitice so when using the functions in this module make sure the inputs you provide are in lowercase to. If you need to convert your input to lowercase use string:casefold/1.

Types

add_error/0

Specs

add_error() :: no_such_realm | reserved_name | already_exists.

external/0

Specs

external() :: t().

list_opts/0

Specs

list_opts() :: #{limit => pos_integer()}.

name/0

Specs

name() :: binary() | anonymous | all.

t/0

Specs

t() ::
    #{type := group,
      version := binary(),
      name := binary() | anonymous,
      groups := [binary()],
      meta => #{binary() => any()}}.

uri/0

Specs

uri() :: binary().

Functions

add(RealmUri, Group)

Specs

add(uri(), t()) -> {ok, t()} | {error, any()}.

add_group(RealmUri, Groups, Groupname)

Specs

add_group(RealmUri :: uri(), Groups :: all | t() | [t()] | name() | [name()], Groupname :: name()) ->
             ok.

Adds group named Groupname to gropus Groups in realm with uri RealmUri.

add_groups(RealmUri, Groups, Groupnames)

Specs

add_groups(RealmUri :: uri(),
           Groups :: all | t() | [t()] | name() | [name()],
           Groupnames :: [name()]) ->
              ok.

Adds groups Groupnames to gropus Groups in realm with uri RealmUri.

add_or_update(RealmUri, Gropu)

Specs

add_or_update(RealmUri :: uri(), Gropu :: t()) -> {ok, t()} | {error, add_error()}.

Adds a new user or updates an existing one. This change is globally replicated.

exists(RealmUri, Name)

Specs

exists(uri(), list() | binary()) -> boolean().

fetch(RealmUri, Name)

Specs

fetch(uri(), list() | binary()) -> t() | no_return().

groups(_)

Specs

groups(t()) -> [name()].

Returns the group names the user User is member of.

is_member(Name, Group)

Specs

is_member(Name :: name(), Group :: t()) -> boolean().

Returns true if group Group is a member of the group named Name. Otherwise returns false.

list(RealmUri)

Specs

list(uri()) -> [t()].

list(RealmUri, Opts)

Specs

list(RealmUri :: uri(), Opts :: list_opts()) -> [t()].

lookup(RealmUri, Name0)

Specs

lookup(uri(), list() | binary()) -> t() | {error, not_found}.

meta(Group)

Specs

meta(Group :: t()) -> map().

Returns the metadata map associated with the group Group.

name(_)

Specs

name(t()) -> name().

Returns the group names the user's username.

new(Data)

Specs

new(Data :: map()) -> Group :: t().

normalise_name(Term)

Specs

normalise_name(Term :: name()) -> name() | no_return().

remove(RealmUri, Name)

Specs

remove(uri(), binary() | map()) -> ok | {error, unknown_group | reserved_name}.

remove_group(RealmUri, Groups, Groupname)

Specs

remove_group(RealmUri :: uri(),
             Groups :: all | t() | [t()] | name() | [name()],
             Groupname :: name()) ->
                ok.

Removes groups Groupnames from gropus Groups in realm with uri RealmUri.

remove_groups(RealmUri, Groups, Groupnames)

Specs

remove_groups(RealmUri :: uri(),
              Groups :: all | t() | [t()] | name() | [name()],
              Groupnames :: [name()]) ->
                 ok.

Removes groups Groupnames from gropus Groups in realm with uri RealmUri.

to_external(Group)

Specs

to_external(Group :: t()) -> external().

Returns the external representation of the user User.

topsort(L)

Specs

topsort([t()]) -> [t()].

Creates a directed graph of the groups Groups by traversing the group membership relationship and computes the topological ordering of the groups if such ordering exists. Otherwise returns Groups unmodified. Fails with {cycle, Path :: [name()]} exception if the graph directed graph has cycles of length two or more.
This function doesn't fetch the definition of the groups in each group groups property.

unknown(RealmUri, Names)

Specs

unknown(RealmUri :: uri(), Names :: [binary()]) -> Unknown :: [binary()].

Takes a list of groupnames and returns any that can't be found on the realm identified by RealmUri or in its prototype (if set).

update(RealmUri, Name, Data)

Specs

update(RealmUri :: uri(), Name :: binary(), Data :: map()) ->
          {ok, NewUser :: t()} | {error, any()}.

Name cannot be a reserved name. See bondy_rbac:is_reserved_name/1.