bondy_rbac_source (bondy v1.0.0-beta.16)

**Note:** Usernames and group names are stored in lower case. All functions in this module are case sensitice so when using the functions in this module make sure the inputs you provide are in lowercase to. If you need to convert your input to lowercase use string:casefold/1.

Types

assignment/0

Specs

assignment() :: #source_assignment{}.

external/0

Specs

external() :: t().

list_opts/0

Specs

list_opts() :: #{limit => pos_integer()}.

t/0

Specs

t() ::
    #{type := source,
      version := binary(),
      username := binary() | all | anonymous,
      cidr := bondy_cidr:t(),
      authmethod := binary(),
      meta => #{binary() => any()}}.

uri/0

Specs

uri() :: binary().

user_source/0

Specs

user_source() ::
    #{type := source,
      version := binary(),
      username := binary() | all | anonymous,
      cidr := bondy_cidr:t(),
      authmethod := binary(),
      meta => #{binary() => any()}}.

Functions

add(RealmUri, Assignment)

Specs

add(RealmUri :: uri(), Assignment :: map() | assignment()) -> ok | {error, any()}.

Adds a source to the realm identified by RealmUri using assignment or map Assignment.

add(Realmuri, Usernames, Assignment)

Specs

add(Realmuri :: uri(),
    Usernames :: [binary()] | all | anonymous,
    Assignment :: map() | assignment()) ->
       ok | {error, any()}.

authmethod(_)

Returns the authmethod associated withe the source

cidr(_)

Returns the source's CIDR.

list(RealmUri)

Specs

list(uri()) -> [t()].

list(RealmUri, Opts)

Specs

list(RealmUri :: uri(), Opts :: list_opts()) -> [t()].

match(RealmUri, Username)

Specs

match(uri(), binary() | all | anonymous) -> [t()].

Returns all the sources for user including the ones for speacial use 'all'.

match(RealmUri, Username, ConnIP)

Specs

match(RealmUri :: uri(), Username :: binary() | all | anonymous, ConnIP :: inet:ip_address()) ->
         [t()].

match_first(RealmUri, Username, ConnIP)

Specs

match_first(RealmUri :: uri(),
            Username :: binary() | all | anonymous,
            ConnIP :: inet:ip_address()) ->
               {ok, t()} | {error, nomatch}.

Returns the first matching source of all the sources available for username Username.

meta(_)

Returns the metadata associated with the source

new_assignment(Data)

Specs

new_assignment(Data :: map()) -> Source :: assignment().

remove(RealmUri, Usernames, CIDR)

Specs

remove(RealmUri :: uri(),
       Usernames :: [binary() | anonymous] | binary() | anonymous | all,
       CIDR :: bondy_cidr:t()) ->
          ok.

remove_all(RealmUri, Username)

Specs

remove_all(RealmUri :: uri(), Username :: binary()) -> ok.

to_external(Source)

Specs

to_external(Source :: t()) -> external().

Returns the external representation of the source Source.