bondy_rbac_user (bondy v1.0.0-beta.16)

A user is a role that is able to log into a Bondy Realm. Users have attributes associated with themelves like username, credentials (password or authorized keys) and metadata determined by the client applications. Users can be assigned group memberships.

Types

add_error/0

Specs

add_error() :: no_such_realm | reserved_name | already_exists.

add_opts/0

Specs

add_opts() :: #{password_opts => bondy_password:opts()}.

external/0

Specs

external() ::
    #{type := user,
      version := binary(),
      username := username(),
      groups := [binary()],
      has_password := boolean(),
      has_authorized_keys := boolean(),
      sso_realm_uri => maybe(uri()),
      meta => #{binary() => any()}}.

list_opts/0

Specs

list_opts() :: #{limit => pos_integer()}.

maybe/1

Specs

maybe(T) :: T | undefined.

new_opts/0

Specs

new_opts() :: #{password_opts => bondy_password:opts()}.

t/0

Specs

t() ::
    #{type := user,
      version := binary(),
      username := username(),
      groups := [binary()],
      password => bondy_password:future() | bondy_password:t(),
      authorized_keys => [binary()],
      sso_realm_uri => maybe(uri()),
      meta => #{binary() => any()},
      password_opts => bondy_password:opts()}.

update_error/0

Specs

update_error() ::
    no_such_realm | reserved_name |
    {no_such_user, username()} |
    {no_such_groups, [bondy_rbac_group:name()]}.

update_opts/0

Specs

update_opts() :: #{update_credentials => boolean(), password_opts => bondy_password:opts()}.

uri/0

Specs

uri() :: binary().

username/0

Specs

username() :: binary() | anonymous.

Functions

add(RealmUri, User)

Specs

add(uri(), t()) -> {ok, t()} | {error, add_error()}.

Adds a new user to the RBAC store. User MUST have been created using new/1 or new/2. This record is globally replicated.

add_group(RealmUri, Users, Groupname)

Specs

add_group(RealmUri :: uri(),
          Users :: all | t() | [t()] | username() | [username()],
          Groupname :: bondy_rbac_group:name()) ->
             ok | {error, Reason :: any()}.

Adds group named Groupname to users Users in realm with uri RealmUri.

add_groups(RealmUri, Users, Groupnames)

Specs

add_groups(RealmUri :: uri(),
           Users :: all | t() | [t()] | username() | [username()],
           Groupnames :: [bondy_rbac_group:name()]) ->
              ok | {error, Reason :: any()}.

Adds groups Groupnames to users Users in realm with uri RealmUri.

add_or_update(RealmUri, User)

Specs

add_or_update(RealmUri :: uri(), User :: t()) -> {ok, t()} | {error, add_error()}.

Adds a new user or updates an existing one. This change is globally replicated.

add_or_update(RealmUri, User, Opts)

Specs

add_or_update(RealmUri :: uri(), User :: t(), Opts :: update_opts()) ->
                 {ok, t()} | {error, add_error()}.

Adds a new user or updates an existing one. This change is globally replicated.

authorized_keys(_)

Returns the list of authorized keys for this user. These keys are used with the WAMP Cryptosign authentication method or equivalent.

change_password(RealmUri, Username, New)

change_password(RealmUri, Username, New, Old)

disable(RealmUri, User)

Specs

disable(RealmUri :: uri(), User :: t()) -> ok | {error, any()}.

Sets the value of the enabled property to false. See is_enabled/2.

enable(RealmUri, User)

Specs

enable(RealmUri :: uri(), User :: t()) -> ok | {error, any()}.

Sets the value of the enabled property to true. See is_enabled/2.

exists(RealmUri, Username)

Specs

exists(RealmUri :: uri(), Username :: binary()) -> boolean().

fetch(RealmUri, Username)

Specs

fetch(uri(), binary()) -> t() | no_return().

groups(_)

Returns the group names the user User is member of.

has_authorized_keys(User)

Specs

has_authorized_keys(User :: t()) -> boolean().

Returns true if user User has a authorized keys. Otherwise returns false. See authorized_keys/1.

has_password(User)

Specs

has_password(User :: t()) -> boolean().

Returns true if user User has a password. Otherwise returns false.

is_enabled(User)

Specs

is_enabled(User :: t()) -> boolean().

Returns true if user User is active. Otherwise returns false. A user that is not active cannot establish a session. See enable/3 and disable/3.

is_enabled(RealmUri, Username)

Specs

is_enabled(RealmUri :: uri(), Username :: binary()) -> boolean().

Returns true if user identified with Username is enabled. Otherwise returns false. A user that is not enabled cannot establish a session. See enable/2 and disable/3.

is_member(Name0, User)

Specs

is_member(Name0 :: bondy_rbac_group:name(), User :: t()) -> boolean().

Returns true if user User is a member of the group named Name. Otherwise returns false.

is_sso_user(User)

Specs

is_sso_user(User :: t()) -> boolean().

Returns true if user User is managed in a SSO Realm, false if it is locally managed.

list(RealmUri)

Specs

list(uri()) -> [t()].

list(RealmUri, Opts)

Specs

list(RealmUri :: uri(), Opts :: list_opts()) -> [t()].

lookup(RealmUri, Username)

Specs

lookup(RealmUri :: uri(), Username :: binary()) -> t() | {error, not_found}.

meta(User)

Specs

meta(User :: t()) -> map().

Returns the metadata map associated with the user User.

new(Data)

Specs

new(Data :: map()) -> User :: t().

new(Data, Opts)

Specs

new(Data :: map(), Opts :: new_opts()) -> User :: t().

normalise_username(Term)

Specs

normalise_username(Term :: username()) -> username() | no_return().

password(User)

Specs

password(User :: t()) -> maybe(bondy_password:future() | bondy_password:t()).

Returns the password object or undefined if the user does not have a password. See bondy_password.

remove(RealmUri, Username0)

Specs

remove(uri(), binary() | map()) -> ok | {error, {no_such_user, username()} | reserved_name}.

remove_group(RealmUri, Users, Groupname)

Specs

remove_group(RealmUri :: uri(),
             Users :: all | t() | [t()] | username() | [username()],
             Groupname :: bondy_rbac_group:name()) ->
                ok.

Removes groups Groupnames from users Users in realm with uri RealmUri.

remove_groups(RealmUri, Users, Groupnames)

Specs

remove_groups(RealmUri :: uri(),
              Users :: all | t() | [t()] | username() | [username()],
              Groupnames :: [bondy_rbac_group:name()]) ->
                 ok.

Removes groups Groupnames from users Users in realm with uri RealmUri.

resolve(User)

Specs

resolve(User :: t()) -> Resolved :: t() | no_return().

If the user User is not sso-managed, returns User unmodified. Otherwise, fetches the user's credentials, the enabled status and additional metadata from the SSO Realm and merges it into User using the following procedure:

* Copies the password and authorized_keys from the SSO user into User. * Adds the meta` contents from the SSO user to a key names `sso to the Usermeta map. * Sets the enabled property by performing the conjunction (logical AND) of both user records.
The call fails with an exception if the SSO user associated with User was not found.

sso_realm_uri(User)

Specs

sso_realm_uri(User :: t()) -> maybe(uri()).

Returns the URI of the Same Sign-on Realm in case the user is a SSO user. Otherwise, returns undefined.

to_external(User)

Specs

to_external(User :: t()) -> external().

Returns the external representation of the user User.

unknown(RealmUri, Usernames)

Specs

unknown(RealmUri :: uri(), Usernames :: [username()]) -> Unknown :: [username()].

Takes a list of usernames and returns any that can't be found.

update(RealmUri, Username, Data)

Specs

update(RealmUri :: uri(), Username :: binary(), Data :: map()) ->
          {ok, NewUser :: t()} | {error, update_error()}.

Updates an existing user. This change is globally replicated.

update(RealmUri, UserOrUsername, Data, Opts)

Specs

update(RealmUri :: uri(), UserOrUsername :: t() | binary(), Data :: map(), Opts :: update_opts()) ->
          {ok, NewUser :: t()} | {error, any()}.

Updates an existing user. This change is globally replicated.

username(_)

Returns the group names the user's username.