bondy_realm (bondy v1.0.0-beta.16)

Realms are routing and administrative domains and act as namespaces for all resources in Bondy i.e. all users, groups, permissions, registrations and subscriptions belong to a Realm. Messages and events are routed separately for each individual realm so sessions attached to a realm won’t see message and events occurring on another realm.

Overview

The realm is a central and fundamental concept in Bondy. It does not only serve as an authentication and authorization domain but also as a message routing domain. Bondy ensures no messages routed in one realm will leak into another realm.

Security

A realm's security may be checked, enabled, or disabled by an administrator through the WAMP and HTTP APIs. This allows an administrator to change security settings of a realm on the whole cluster quickly without needing to change settings on a node-by-node basis.

If you disable security, this means that you have disabled all of the various authentication and authorization checks that take place when establishing a session and executing operations against a Bondy Realm. Users, groups, and other security resources remain available for configuration while security is disabled, and will be applied if and when security is re-enabled.

Realm security is enabled by default.

Storage

Realms (and the associated users, credentials, groups, sources and permissions) are persisted to disk and replicated across the cluster using the plum_db subsystem.

Bondy Master Realm

When you start Bondy for the first time it creates and stores the Bondy Master realm a.k.a com.leapsight.bondy. This realm is the root realm which allows an admin user to create, list, modify and delete other realms.

Realm Properties

uriuri()[required, immutable]
The realm identifier.

descriptionbinary()
A textual description of the realm.

is_prototypeboolean()' [immutable] Iftrue' this realm is a realm used as a prototype.
Default: false

prototype_uriuri()
If present, this it the URI of the the realm prototype this realm inherits some of its behaviour and features from.

sso_realm_uriuri()
If present, this it the URI of the SSO Realm this realm is connected to.

is_sso_realmboolean()' [immutable] Iftrue' this realm is an SSO Realm.
Default: false.

allow_connectionsboolean()' Iftrue' this realm is allowing connections from clients. It is normally set to false' when the realm is an SSO Realm. Default:true'

authmethodslist(binary()' The list of the authentication methods allowed by this realm. Default:[anonymous, password, ticket, oauth2, wampcra]'

security_statusbinary()' The stringenabled' if security is enabled. Otherwise the string disabled.

public_keyslist()
A list of JWK values.

Realm Prototypes

A Prototype Realm is a realm that acts as a prototype for the construction of other realms. A prototype realm is a normal realm whose property is_prototype has been set to true.

Prototypical inheritance allows us to reuse the properties (including RBAC definitions) from one realm to another through a reference URI configured on the prototype_uri property.

Prototypical inheritance is a form of single inheritance as realms are can only be related to a single prototype.

The prototype_uri property is defined as an *irreflexive property* i.e. a realm cannot have itself as prototype. In addition *a prototype cannot inherit from another prototype*. This means the inheritance chain is bounded to one level.

Inherited properties

The following is the list of properties which a realm inherits from a prototype when those properties have not been asigned a value. Setting a value to these properties is equivalente to overriding the prototype's.

* **security_enabled** * **allow_connections** * **sso_realm_uri** * **authmethods**

In addition realms inherit Groups, Sources and Grants from their prototype. The following are the inheritance rules:

1. Users cannot be defined at the prototype i.e. no user inheritance. 1. A realm has access to all groups defined in the prototype i.e. from a realm perspective the prototype groups operate in the same way as if they have been defined in the realm itself. This enables roles (users and groups) in a realm to be members of groups defined in the prototype. 1. A group defined in a realm overrides any homonymous group in the prototype. This works at all levels of the group membership chain. 1. The previous rule does not apply to the special group all. Permissions granted to all are merged between a realm and its prototype.

Same Sign-on (SSO)

Bondy SSO (Same Sign-on) is a feature that allows users to access multiple realms using just one set of credentials.

It is enabled by setting the realm's sso_realm_uri property during realm creation or during an update operation.
* It requires the user to authenticate when opening a session in a realm. * Changing credentials e.g. updating password can be performed while connected to any realm.

Types

external/0

Specs

external() ::
    #{uri := uri(),
      is_prototype := boolean(),
      prototype_uri := maybe(uri()),
      description := binary(),
      authmethods := [binary()],
      is_sso_realm := boolean(),
      allow_connections := boolean(),
      public_keys := keyset(),
      security_status := enabled | disabled}.

keyset/0

Specs

keyset() :: [map()].

maybe/1

Specs

maybe(T) :: T | undefined.

t/0
(opaque)

Specs

t()

uri/0

Specs

uri() :: binary().

Functions

allow_connections(Realm)

Specs

allow_connections(Realm :: t() | uri()) -> boolean().

Returns true if the Realm is allowing connections. Otherwise returns false.

If the value is undefined and the realm has a prototype the prototype's value is returned. Otherwise if the realm doesn't have a prototype returns false.

Note that a Prototype realm never allows connections irrespective of the value set to this property. This this property is just used as a template for realms to inherit from.
This setting is used to either temporarilly restrict new connections to the realm or to avoid connections when the realm is used as a Single Sign-on Realm. When connections are not allowed the only way of managing the resources in the realm is through a connection to the Bondy Master Realm.

apply_config()

Specs

apply_config() -> ok | no_return().

Loads a security config file from bondy_config:get([security, config_file]) if defined and applies its definitions.

authmethods(Realm)

Specs

authmethods(Realm :: t() | uri()) -> [binary()].

Returns the list of supported authentication methods for Realm.

If the value is undefined and the realm has a prototype the prototype's value is returned. Otherwise if the realm doesn't have a prototype returns the default list of authentication methods.
See is_allowed_authmethod for more information about how this affects the methods available for an authenticating user.

create(Map0)

Specs

create(uri() | map()) -> t() | no_return().

delete(Realm)

Specs

delete(t() | uri()) -> ok | {error, not_found | active_users} | no_return().

description(Realm)

Specs

description(t() | uri()) -> map().

disable_security(Realm)

Specs

disable_security(t() | uri()) -> ok | no_return().

Disables security for realm Realm.

enable_security(Realm)

Specs

enable_security(t() | uri()) -> ok.

Enables security for realm Realm.

encryption_keys(Realm)

Specs

encryption_keys(t() | uri()) -> [map()].

exists(Uri)

Specs

exists(uri()) -> boolean().

fetch(Uri)

Specs

fetch(uri()) -> t().

Retrieves the realm identified by Uri from the tuplespace. If the realm does not exist it fails with reason '{badarg, Uri}'.

from_file(Filename)

Specs

from_file(Filename :: file:filename_all()) -> ok | no_return().

Loads a security config file from Filename.

get(Uri)

Specs

get(uri()) -> t() | {error, not_found}.

Retrieves the realm identified by Uri from the tuplespace. If the realm does not exist and automatic creation of realms is enabled, it will add a new one for Uri with the default configuration options.

get(Uri, Opts)

Specs

get(uri(), map()) -> t() | {error, not_found}.

Retrieves the realm identified by Uri from the tuplespace. If the realm does not exist and automatic creation of realms is enabled, it will create a new one for Uri with configuration options Opts.

get_encryption_key(Realm, Kid)

Specs

get_encryption_key(t() | uri(), Kid :: binary()) -> map() | undefined.

get_private_key(Realm, Kid)

Specs

get_private_key(t() | uri(), Kid :: binary()) -> map() | undefined.

get_public_key(Realm, Kid)

Specs

get_public_key(t() | uri(), Kid :: binary()) -> map() | undefined.

get_random_encryption_kid(Realm)

Specs

get_random_encryption_kid(t() | uri()) -> map().

get_random_kid(Realm)

Specs

get_random_kid(t() | uri()) -> binary().

grants(Realm)

Specs

grants(Realm :: t() | uri()) -> [bondy_rbac_user:t()].

Returns the list of grants belonging to realm Realm. These includes the grants inherited from the prototype (if defined).

grants(Realm, Opts)

Specs

grants(Realm :: t() | uri(), Opts :: map()) -> [bondy_rbac_user:t()].

Returns the list of grants belonging to realm Realm. These includes the grants inherited from the prototype (if defined).

groups(Realm)

Specs

groups(Realm :: t() | uri()) -> [bondy_rbac_user:t()].

Returns the list of users belonging to realm Realm. These includes the groups inherited from the prototype (if defined).

groups(Realm, Opts)

Specs

groups(Realm :: t() | uri(), Opts :: map()) -> [bondy_rbac_user:t()].

Returns the list of groups belonging to realm Realm. These includes the groups inherited from the prototype (if defined).

info(Realm)

Specs

info(t() | uri()) -> map().

is_allowed_authmethod(Realm, Method)

Specs

is_allowed_authmethod(Realm :: t() | uri(), Method :: binary()) -> boolean().

Returns true if Method is an authentication method supported by realm Realm. Otherwise returns false.

The fact that method Method is included in the realm's authmethods (See {3link authmethods/1}) is no guarantee that the method will be available for a particular user.
The availability is also affected by the source rules defined for the realm and the capabilities of each user e.g. if the user has no password then the password-based authentication methods in this list will not be available.

is_allowed_sso_realm(Realm, SSORealmUri)

Specs

is_allowed_sso_realm(Realm :: t() | uri(), SSORealmUri :: uri()) -> boolean().

Returns true if realm Realm is associated with the SSO Realm identified by uri SSORealmUri`. Otherwise returns `false.

is_prototype(Realm)

Specs

is_prototype(Realm :: t() | uri()) -> boolean().

Returns true if realm Realm is a prototype. Otherwise, returns false.

**Pre-conditions** * The property prototype_uri MUST be undefined. * This property cannot be set to false once it has been set to true.
**Post-conditions** * If this property is true, the prototype_uri cannot be set.

is_security_enabled(Realm)

Specs

is_security_enabled(t() | uri()) -> boolean().

Returns true if security is enabled. Otherwise returns false.

If the value is undefined and the realm has a prototype the prototype's value is returned. Otherwise if the realm doesn't have a prototype returns true (default).

Security for this realm can be enabled or disabled using the functions enable_security/1 and disable_security/1 respectively.
See security_status/1 if you want the security status representation as an atom.

is_sso_realm(Realm)

Specs

is_sso_realm(Realm :: t() | uri()) -> boolean().

Returns true if the Realm is a Same Sign-on (SSO) realm. Otherwise returns false.
If the value is undefined and the realm has a prototype the prototype's value is returned. Otherwise if the realm doesn't have a prototype returns false.

is_value_inherited(Realm, Property)

Specs

is_value_inherited(Realm :: t() | uri(), Property :: atom()) -> boolean() | no_return().

Returns true if the property value is inherited from a prototype. Otherwise returns false.

list()

Specs

list() -> [t()].

lookup(Uri)

Specs

lookup(uri()) -> t() | {error, not_found}.

Retrieves the realm identified by Uri from the tuplespace or '{error, not_found}' if it doesn't exist.

password_opts(Realm)

Specs

password_opts(t() | uri()) -> maybe(bondy_password:opts()).

Returns the password options to be used as default when adding users to this realm. If the options have not been defined returns atom undefined.

private_keys(Realm)

Specs

private_keys(t() | uri()) -> [map()].

prototype_uri(Realm)

Specs

prototype_uri(Realm :: t() | uri()) -> maybe(uri()).

Returns the uri of realm Realm prototype if defined. Otherwise returns undefined.

public_keys(Realm)

Specs

public_keys(t() | uri()) -> [map()].

security_status(Term)

Specs

security_status(t() | uri()) -> enabled | disabled.

A util function that returns the security status as an atom. See is_security_enabled/1.

sources(Realm)

Specs

sources(Realm :: t() | uri()) -> [bondy_rbac_user:t()].

Returns the list of sources belonging to realm Realm. These includes the sources inherited from the prototype (if defined).

sources(Realm, Opts)

Specs

sources(Realm :: t() | uri(), Opts :: map()) -> [bondy_rbac_user:t()].

Returns the list of sources belonging to realm Realm. These includes the sources inherited from the prototype (if defined).

sso_realm_uri(Realm)

Specs

sso_realm_uri(Realm :: t() | uri()) -> maybe(uri()).

Returns the same sign on (SSO) realm URI used by the realm.

If a value is set, then all authentication and user creation will be done on the the SSO Realm.

Groups, permissions and sources are still managed by this realm (or the prototype it inherits from).
If the value is undefined and the realm has a prototype the prototype's value is returned. Otherwise if the realm doesn't have a prototype returns undefined.

to_external(Realm)

Specs

to_external(t() | uri()) -> external().

Returns the external map representation of the realm.

update(Realm, Data)

Specs

update(Realm :: t() | uri(), Data :: map()) -> Realm :: t() | no_return().

uri(Realm)

Specs

uri(Realm :: t()) -> uri().

Returns the URI that identifies the realm Realm.

users(Realm)

Specs

users(Realm :: t() | uri()) -> [bondy_rbac_user:t()].

Returns the list of users belonging to realm Realm. Users are never inherited through prototypes.

users(Realm, Opts)

Specs

users(Realm :: t() | uri(), Opts :: map()) -> [bondy_rbac_user:t()].

Returns the list of users belonging to realm Realm. Users are never inherited through prototypes.