bondy_ref (bondy v1.0.0-beta.16)

A bondy_ref (reference) acts as a fully qualified name for a process or callback function in a Bondy network. The reference is used by Bondy in the bondy_registry when registering procedures and subscriptions so that the bondy_router can forward and/or relay a message to a process, or callback function.

## Types ### Internal

### Relay

### Client
### Callback

Types

bridge_relay/0

Specs

bridge_relay() :: #bondy_ref{type :: bridge_relay}.

client/0

Specs

client() :: #bondy_ref{type :: client}.

id/0

Specs

id() :: 0..9007199254740992.

internal/0

Specs

internal() :: #bondy_ref{type :: internal}.

maybe/1

Specs

maybe(T) :: T | undefined.

mf/0

Specs

mf() :: {M :: module(), F :: atom()}.

name/0

Specs

name() :: term().

nodestring/0

Specs

nodestring() :: binary().

ref_type/0

Specs

ref_type() :: relay | bridge_relay | internal | client.

relay/0

Specs

relay() :: #bondy_ref{type :: relay}.

t/0

Specs

t() :: #bondy_ref{}.

target/0

Specs

target() :: {pid, binary()} | {name, term()} | {callback, mf()}.

target_type/0

Specs

target_type() :: pid | name | callback.

uri/0

Specs

uri() :: binary().

wildcard/1

Specs

wildcard(T) :: T | '_'.

Functions

callback(Bondy_ref)

Specs

callback(t()) -> maybe(mf()).

is_bridge_relay(Ref)

Specs

is_bridge_relay(t()) -> boolean().

is_client(Ref)

Specs

is_client(t()) -> boolean().

is_internal(Ref)

Specs

is_internal(t()) -> boolean().

is_local(Ref)

Specs

is_local(Ref :: t()) -> boolean().

Returns whether this is a reference to a local target i.e. located on the caller's node.

is_relay(Ref)

Specs

is_relay(t()) -> boolean().

is_self(Ref)

Specs

is_self(Ref :: t()) -> boolean().

Returns whether this is a reference to the calling process.

is_type(Bondy_ref)

Specs

is_type(t()) -> boolean().

Returns true if term Term is a reference. Otherwise returns false.

name(Bondy_ref)

Specs

name(t()) -> maybe(term()).

new(Type, RealmUri)

Specs

new(Type :: ref_type(), RealmUri :: uri()) -> t().

new(Type, RealmUri, Target)

Specs

new(Type :: ref_type(), RealmUri :: uri(), Target :: pid() | mf() | name()) -> t().

new(Type, RealmUri, Target, SessionId)

Specs

new(Type :: ref_type(),
    RealmUri :: uri(),
    Target :: pid() | mf() | name(),
    SessionId :: maybe(bondy_session:id())) ->
       t().

new(Type, RealmUri, Target, SessionId, Node)

Specs

new(Type :: ref_type(),
    RealmUri :: uri(),
    Target :: pid() | mf() | name(),
    SessionId :: maybe(bondy_session:id()),
    Node :: node() | nodestring()) ->
       t().

Creates a new reference.

node(Bondy_ref)

Specs

node(t()) -> node().

Returns the Bondy peer node in which the target of this reference is located and/or connected to.
See target/1 for a description of the different targets and the relationship with the node.

nodestring(Bondy_ref)

Specs

nodestring(t()) -> nodestring().

Returns the Bondy peer binary string name of the node in which the target of this reference is located and/or connected to.
See target/1 for a description of the different targets and the relationship with the node.

pattern(Type, RealmUri, Target, SessionId, Node)

Specs

pattern(Type :: wildcard(ref_type()),
        RealmUri :: uri(),
        Target :: wildcard(pid() | mf() | name()),
        SessionId :: wildcard(maybe(bondy_session:id())),
        Node :: wildcard(node() | nodestring())) ->
           t().

pid(Bondy_ref)

Specs

pid(t()) -> maybe(pid()) | no_return().

Returns a pid when target is of type {pid, binary()}, otherwise returns the atom undefined. If the reference is not local it fails with exception not_my_node. This is becuase process identifies can only be used on the node where they were created (this is because we are using Partisan and not Distributed Erlang).

realm_uri(Bondy_ref)

Specs

realm_uri(t()) -> uri().

Returns the realm URI of the reference.

session_id(Bondy_ref)

Specs

session_id(t()) -> maybe(id()).

Returns the session identifier of the reference or the atom undefined.

target(Bondy_ref)

Specs

target(t()) -> target().

Returns the target of the reference. A target is a process (pid()) a gproc registered name ({name, GprocName}) or a callback (mf()).

If the reference type is client the target refers to the session and/or process owning the network connection to the client.

If the reference type is internal the target refers to an internal process acting as one of the Bondy client roles e.g. an internal subscriber. Messages destined to a target located on a different node are relayed by the router through the cluster distribution layer (Partisan).
If the reference type is relay the target refers to the session and/or process owning the network connection to the edge or remote cluster (See bondy_edge_uplink_client and bondy_edge_uplink_server respectively). The relay acts as a proxy for the actual edge or remote clients, internal and callback targets.

target_type(Bondy_ref)

Specs

target_type(t()) -> wildcard(target_type()).

Returns the target of the reference. A target is a process (pid()) a gproc registered name ({name, GprocName}) or a callback (mf()).

If the reference type is client the target refers to the session and/or process owning the network connection to the client.

If the reference type is internal the target refers to an internal process acting as one of the Bondy client roles e.g. an internal subscriber. Messages destined to a target located on a different node are relayed by the router through the cluster distribution layer (Partisan).
If the reference type is relay the target refers to the session and/or process owning the network connection to the edge or remote cluster (See bondy_edge_uplink_client and bondy_edge_uplink_server respectively). The relay acts as a proxy for the actual edge or remote clients, internal and callback targets.

type(Term)

Specs

type(Term :: term()) -> ref_type().

Returns the reference type.

types()

Specs

types() -> [ref_type()].

Returns all the supported reference types.