bondy_registry (bondy v1.0.0-beta.16)

An in-memory registry for PubSub subscriptions and Routed RPC registrations, providing pattern matching capabilities including support for WAMP's version 2.0 match policies (exact, prefix and wildcard).

The registry is stored both in an in-memory distributed table (plum_db). Also an in-memory trie-based indexed (materialised vieq) is used for exact and prefix matching.

Note: support for wildcard matching is soon to be supported.
This module also provides a singleton server to perform the initialisation of the trie from the plum_db tables.

Types

continuation/0

Specs

continuation() :: {bondy_registry_entry:entry_type(), plum_db:continuation()}.

continuation_or_eot/0

Specs

continuation_or_eot() :: continuation_or_eot().

eot/0

Specs

eot() :: '$end_of_table'.

id/0

Specs

id() :: 0..9007199254740992.

task/0

Specs

task() :: fun((bondy_registry_entry:t(), bondy_context:t()) -> ok).

uri/0

Specs

uri() :: binary().

Functions

add(Entry)

Specs

add(bondy_registry_entry:t()) ->
       {ok, IsFirstEntry :: boolean()} | {error, already_exists} | no_return().

Adds an existing entry to the registry.

Adding an already existing entry is treated differently based on whether the entry is a registration or a subscription.

According to the WAMP specification, in the case of a subscription that was already added before by the same _Subscriber_, the _Broker_ should not fail and answer with a "SUBSCRIBED" message, containing the existing "Subscription|id". So in this case this function returns {ok, bondy_registry_entry:t(), boolean()}.

In case of a registration, as a default, only a single Callee may register a procedure for an URI. However, when shared registrations are supported, then the first Callee to register a procedure for a particular URI MAY determine that additional registrations for this URI are allowed, and what Invocation Rules to apply in case such additional registrations are made.
This is configured through the 'invoke' options. When invoke is not 'single', Dealer MUST fail all subsequent attempts to register a procedure for the URI where the value for the invoke option does not match that of the initial registration. Accordingly this function might return an error tuple.

add(Type, RegUri, Opts, CtxtOrRef)

Specs

add(Type :: bondy_registry_entry:entry_type(),
    RegUri :: uri(),
    Opts :: map(),
    CtxtOrRef :: bondy_context:t() | bondy_ref:t()) ->
       {ok, Entry :: bondy_registry_entry:t(), IsFirstEntry :: boolean()} |
       {error, {already_exists, bondy_registry_entry:t()} | any()}.

Adds an entry to the registry.

Adding an already existing entry is treated differently based on whether the entry is a registration or a subscription.

According to the WAMP specification, in the case of a subscription that was already added before by the same _Subscriber_, the _Broker_ should not fail and answer with a "SUBSCRIBED" message, containing the existing "Subscription|id". So in this case this function returns {ok, bondy_registry_entry:t(), boolean()}.

In case of a registration, as a default, only a single Callee may register a procedure for an URI. However, when shared registrations are supported, then the first Callee to register a procedure for a particular URI MAY determine that additional registrations for this URI are allowed, and what Invocation Rules to apply in case such additional registrations are made.
This is configured through the 'invoke' options. When invoke is not 'single', Dealer MUST fail all subsequent attempts to register a procedure for the URI where the value for the invoke option does not match that of the initial registration. Accordingly this function might return an error tuple.

code_change(OldVsn, State, Extra)

entries(_)

Specs

entries(continuation()) -> {[bondy_registry_entry:t()], continuation_or_eot()} | eot().

Continues returning the list of entries owned by a session started with entries/4.

The next chunk of the size specified in the initial entries/4 call is returned together with a new Continuation, which can be used in subsequent calls to this function.
When there are no more objects in the table, {[], '$end_of_table'} is returned.

entries(Type, Ctxt)

Specs

entries(bondy_registry_entry:entry_type(), bondy_context:t()) -> [bondy_registry_entry:t()].

Returns the list of entries owned by the active session.
This function is equivalent to calling entries/2 with the RealmUri and SessionId extracted from the Context.

entries(Type, RealmUri, Node, SessionId)

Specs

entries(bondy_registry_entry:entry_type(), RealmUri :: uri(), Node :: atom(), SessionId :: id()) ->
           [bondy_registry_entry:t()].

Returns the complete list of entries owned by a session matching RealmUri and SessionId.
Use entries/3 and entries/1 to limit the number of entries returned.

entries(Type, Realm, Node, SessionId, Limit)

Specs

entries(bondy_registry_entry:entry_type(),
        Realm :: uri(),
        Node :: atom() | '_',
        SessionId :: id() | '_',
        Limit :: pos_integer() | infinity) ->
           [bondy_registry_entry:t()] |
           {[bondy_registry_entry:t()], continuation_or_eot()} |
           eot().

Works like entries/3, but only returns a limited (Limit) number of entries. Term Continuation can then be used in subsequent calls to entries/1 to get the next chunk of entries.

handle_call(Event, From, State0)

handle_cast(Event, State)

handle_info(Info, State)

info()

Returns information about the registry

info(_)

init(_)

init_tries()

lookup(Key)

Specs

lookup(Key :: bondy_registry_entry:key()) -> bondy_registry_entry:t() | {error, not_found}.

lookup(Type, EntryId, RealmUri)

lookup(Type, EntryId, RealmUri, Details)

match(_)

Specs

match(continuation()) -> {[bondy_registry_entry:t()], continuation()} | eot().

match(Type, Uri, RealmUri)

Specs

match(bondy_registry_entry:entry_type(), uri(), RealmUri :: uri()) ->
         {[bondy_registry_entry:t()], continuation_or_eot()} | eot().

Calls match/4.

match(Type, Uri, RealmUri, Opts)

Specs

match(bondy_registry_entry:entry_type(), uri(), RealmUri :: uri(), map()) ->
         {[bondy_registry_entry:t()], continuation_or_eot()} | eot().

Returns the entries matching either a topic or procedure Uri according to each entry's configured match specification.
This function is used by the Broker to return all subscriptions that match a topic. And in case of registrations it is used by the Dealer to return all registrations matching a procedure.

remove(Entry)

Specs

remove(bondy_registry_entry:t()) -> ok.

remove(Type, EntryId, Ctxt)

Specs

remove(bondy_registry_entry:entry_type(), id(), bondy_context:t()) -> ok | {error, not_found}.

remove(Type, EntryId, Ctxt, Task)

Specs

remove(bondy_registry_entry:entry_type(), id(), bondy_context:t(), task() | undefined) -> ok.

remove_all(Type, Ctxt)

Specs

remove_all(bondy_registry_entry:entry_type(), bondy_context:t()) -> ok.

Removes all entries matching the context's realm and session_id (if any).

remove_all(Type, Ctxt, Task)

Specs

remove_all(bondy_registry_entry:entry_type(), bondy_context:t(), task() | undefined) -> ok.

Removes all entries matching the context's realm and session_id (if any).

remove_all(Type, RealmUri, Node, SessionId)

Specs

remove_all(bondy_registry_entry:entry_type(),
           RealmUri :: uri(),
           Node :: atom(),
           SessionId :: id()) ->
              [bondy_registry_entry:t()].

Removes all registry entries of type Type, for a {RealmUri, Node SessionId} relation.

start_link()

Starts the registry server. The server maintains the in-memory tries we use for matching and it is also a subscriber for plum_db broadcast and AAE events in order to keep the trie up-to-date with plum_db.

terminate(Reason, State)