bondy_registry_entry (bondy v1.0.0-beta.16)

An entry is a record of a RPC registration or PubSub subscription. It is stored in-memory in the Registry bondy_registry and replicated globally and is immutable.

Types

details_map/0

Specs

details_map() :: #{id => id(), created => calendar:date(), uri => uri(), match => binary()}.

entry_type/0

Specs

entry_type() :: registration | subscription.

external/0

Specs

external() ::
    #{type := entry_type(),
      entry_id := id(),
      uri := uri(),
      match_policy := binary(),
      ref := bondy_ref:t(),
      callback_args := [term()],
      created := calendar:date(),
      options := map(),
      origin_id := maybe(id()),
      origin_ref := maybe(bondy_ref:t())}.

id/0

Specs

id() :: 0..9007199254740992.

key/0

Specs

key() :: #entry_key{}.

maybe/1

Specs

maybe(T) :: T | undefined.

mfargs/0

Specs

mfargs() :: {M :: module(), F :: atom(), A :: maybe([term()])}.

nodestring/0

Specs

nodestring() :: binary().

t/0
(opaque)

Specs

t()

t_or_key/0

Specs

t_or_key() :: t_or_key().

uri/0

Specs

uri() :: binary().

wildcard/1

Specs

wildcard(T) :: T | '_'.

Functions

callback(Entry)

Specs

callback(t()) -> wildcard(maybe(mfargs())).

Returns the callback module when target is a callback, otherwise returns undefined or the wilcard value '_' when the entry was used as a pattern (See pattern/5).

callback_args(Entry)

Specs

callback_args(t()) -> maybe([term()]).

Returns the callback arguments when target is a callback, otherwise returns undefined or the wilcard value '_' when the entry was used as a pattern (See pattern/5).

created(Entry)

Specs

created(t()) -> pos_integer().

Returns the time when this entry was created. Its value is a timestamp in seconds.

get_option(Entry, Key, Default)

Specs

get_option(t(), any(), any()) -> any().

id(Entry)

Specs

id(t_or_key()) -> wildcard(id()).

Returns the value of the subscription's or registration's id property.

is_callback(Entry_key)

Specs

is_callback(t()) -> boolean().

Returns true if the entry target is a callback registration. Callback registrations are only used by Bondy itself to provide some of the admin and meta APIs.

is_entry(Entry)

is_key(Entry_key)

is_local(Entry)

Specs

is_local(t_or_key()) -> boolean().

Returns true if the entry represents a handler local to the caller's node and false when the target is located in a cluster peer.

is_proxy(Entry)

Specs

is_proxy(Entry :: t_or_key()) -> boolean().

key(Entry)

Specs

key(t()) -> key().

Returns the value of the subscription's or registration's realm_uri property.

key_field(_)

key_pattern(Type, Ref)

Specs

key_pattern(entry_type(), Ref :: bondy_ref:t()) -> key().

key_pattern(Type, RealmUri, Node, Extra)

match_policy(Entry)

Specs

match_policy(t()) -> binary().

Returns the match_policy used by this subscription or regitration.

new(Type, Ref, Uri, Options)

Specs

new(entry_type(), bondy_ref:t(), uri(), map()) -> t().

new(Type, RegId, Ref, Uri, Opts0)

Specs

new(entry_type(), id(), bondy_ref:t(), uri(), map()) -> t().

node(Entry)

Specs

node(t_or_key()) -> atom().

Returns the value of the subscription's or registration's node property. This is always the Bondy cluster peer node where the handler exists.

nodestring(Entry)

Specs

nodestring(t_or_key()) -> nodestring().

Returns the value of the subscription's or registration's node property. This is always the Bondy cluster peer node where the handler exists.

options(Entry)

Specs

options(t()) -> map().

Returns the value of the 'options' property of the entry.

origin_id(Entry)

Specs

origin_id(t()) -> maybe(id()).

Returns the origin ref() of the subscription or registration. This value is only present when the entry is a proxy.

origin_ref(Entry)

Specs

origin_ref(t()) -> maybe(bondy_ref:t()).

Returns the origin ref() of the subscription or registration This value is only present when the entry is a proxy.

pattern(Type, RealmUri, ProcedureOrTopic, Options)

Specs

pattern(Type :: entry_type(), RealmUri :: uri(), ProcedureOrTopic :: uri(), Options :: map()) ->
           t().

pattern(Type, RealmUri, ProcedureOrTopic, Options, Extra)

Specs

pattern(Type :: entry_type(),
        RealmUri :: uri(),
        ProcedureOrTopic :: uri(),
        Options :: map(),
        Extra :: map()) ->
           t().

pid(Entry)

Specs

pid(t_or_key()) -> wildcard(maybe(pid())).

Returns the value of the subscription's or registration's pid property when target is a pid() or a session identifier. Otherwise returns undefined or the wilcard value '_' when the entry was used as a pattern (See pattern/5).

proxy(Ref, Entry)

Specs

proxy(Ref :: bondy_ref:bridge_relay(), Entry :: external()) -> t().

Returns a copy of the entry where the node component of the ref has been replaced with the node of the calling process, the session identifier replaced with SessionId and the pid with Pid.
The entry is used by a router node to create a proxy of an entry originated in an edge router.

proxy_details(Entry)

Specs

proxy_details(t()) -> map().

realm_uri(Entry)

Specs

realm_uri(t_or_key()) -> uri().

Returns the value of the subscription's or registration's realm_uri property.

ref(Entry)

Specs

ref(t()) -> bondy_ref:t().

Returns the ref() of the subscription or registration

session_id(Entry)

Specs

session_id(t_or_key()) -> wildcard(maybe(id())).

Returns the value of the subscription's or registration's session_id property.

target(Entry)

Specs

target(t()) -> wildcard(bondy_ref:target()).

Returns either a session identifier, a pid() or a callback module depending on the type of entry. I can also return the wilcard '_' when the entry is used as a pattern (See pattern/5).

to_details_map(Entry)

Specs

to_details_map(t()) -> details_map().

Converts the entry into a map according to the WAMP protocol Details dictionary format.

to_external(Entry)

Specs

to_external(t()) -> external().

Converts the entry into a map. Certain values of type atom such as node are turned into binaries. This is to avoid exhausting a remote node atoms table for use cases where the entries are replicated e.g. edge-remote connections.

type(Entry)

Specs

type(t_or_key()) -> entry_type().

Returns the type of the entry, the atom 'registration' or 'subscription'.

uri(Entry)

Specs

uri(t()) -> uri().

Returns the uri this entry is about i.e. either a subscription topic_uri or a registration procedure_uri.