bondy_retained_message (bondy v1.0.0-beta.16)

This is experimental and does not scale with high traffic at the moment.

Types

continuation/0

Specs

continuation() :: #bondy_retained_continuation{}.

eot/0

Specs

eot() :: '$end_of_table'.

evict_fun/0

Specs

evict_fun() :: fun((uri(), t()) -> ok).

id/0

Specs

id() :: 0..9007199254740992.

match_opts/0

Specs

match_opts() :: #{eligible => [id()], exclude => [id()]}.

t/0

Specs

t() :: #bondy_retained_message{}.

uri/0

Specs

uri() :: binary().

wamp_event/0

Specs

wamp_event() :: #event{}.

Functions

evict_expired()

Specs

evict_expired() -> non_neg_integer().

Evict expired retained messages from all realms.

evict_expired(Realm)

Specs

evict_expired(uri() | '_') -> non_neg_integer().

Evict expired retained messages from realm Realm.

evict_expired(Realm, EvictFun)

Specs

evict_expired(uri() | '_', evict_fun() | undefined) -> non_neg_integer().

Evict expired retained messages from realm Realm or all realms if wildcard '_' is used. Evaluates function Fun for each entry passing Realm and Entry as arguments.

get(Realm, Topic)

Specs

get(Realm :: uri(), Topic :: uri()) -> t() | undefined.

match(Bondy_retained_continuation)

Specs

match(continuation()) -> {[t()] | continuation()} | eot().

match(Realm, Topic, SessionId, Strategy)

Specs

match(Realm :: uri(), Topic :: uri(), SessionId :: id(), Strategy :: binary()) ->
         {[t()], continuation()} | eot().

match(Realm, Topic, SessionId, Strategy, Opts)

Specs

match(Realm :: uri(),
      Topic :: uri(),
      SessionId :: id(),
      Strategy :: binary(),
      Opts :: plum_db:fold_opts()) ->
         {[t()], continuation()} | eot().

put(Realm, Topic, Event, MatchOpts)

Specs

put(Realm :: uri(), Topic :: uri(), Event :: wamp_event(), MatchOpts :: match_opts()) -> ok.

put(Realm, Topic, Event, MatchOpts, TTL)

Specs

put(Realm :: uri(),
    Topic :: uri(),
    Event :: wamp_event(),
    MatchOpts :: match_opts(),
    TTL :: non_neg_integer()) ->
       ok.

size(Mssg)

Specs

size(t()) -> integer().

take(Realm, Topic)

Specs

take(Realm :: uri(), Topic :: uri()) -> t() | undefined.

to_event(Retained, SubscriptionId)

Specs

to_event(Retained :: t(), SubscriptionId :: id()) -> wamp_event().