bondy_retained_message_manager (bondy v1.0.0-beta.16)

Implements Eviction amogst other things

Types

id/0

Specs

id() :: 0..9007199254740992.

uri/0

Specs

uri() :: binary().

wamp_event/0

Specs

wamp_event() :: #event{}.

Functions

code_change(OldVsn, State, Extra)

counters(Realm)

Specs

counters(Realm :: uri()) -> #{messages => integer(), memory => integer()}.

decr_counters(Realm, N, Size)

default_ttl()

Default TTL for retained messages.

get(Realm, Topic)

Specs

get(Realm :: uri(), Topic :: uri()) -> bondy_retained_message:t() | undefined.

handle_call(Event, From, State)

handle_cast(Event, State)

handle_info(Info, State)

incr_counters(Realm, N, Size)

init(_)

match(Cont)

Specs

match(bondy_retained_message:continuation()) ->
         {[bondy_retained_message:t()] | bondy_retained_message:continuation()} |
         bondy_retained_message:eot().

match(Realm, Topic, SessionId, Strategy)

Specs

match(Realm :: uri(), Topic :: uri(), SessionId :: id(), Strategy :: binary()) ->
         {[bondy_retained_message:t()], bondy_retained_message:continuation()} |
         bondy_retained_message:eot().

match(Realm, Topic, SessionId, Strategy, Opts)

Specs

match(Realm :: uri(),
      Topic :: uri(),
      SessionId :: id(),
      Strategy :: binary(),
      Opts :: plum_db:fold_opts()) ->
         {[bondy_retained_message:t()], bondy_retained_message:continuation()} |
         bondy_retained_message:eot().

max_memory()

Maximum space in memory used by retained messages. Once the max has been reached no more events will be stored. A value of 0 means no limit is enforced.

max_message_size()

The max size for an event message. All events whose size exceeds this value will not be retained.

max_messages()

Maximum number of messages that can be store in a Bondy node. Once the max has been reached no more events will be stored. A value of 0 means no limit is enforced.

put(Realm, Topic, Event, MatchOpts)

Specs

put(Realm :: uri(),
    Topic :: uri(),
    Event :: wamp_event(),
    MatchOpts :: bondy_retained_message:match_opts()) ->
       ok.

put(Realm, Topic, Event, MatchOpts, TTL)

Specs

put(Realm :: uri(),
    Topic :: uri(),
    Event :: wamp_event(),
    MatchOpts :: bondy_retained_message:match_opts(),
    TTL :: non_neg_integer() | undefined) ->
       ok.

start_link()

take(Realm, Topic)

Specs

take(Realm :: uri(), Topic :: uri()) -> bondy_retained_message:t() | undefined.

terminate(Reason, State)