bondy_router (bondy v1.0.0-beta.16)

bondy_router provides the routing logic for all interactions.

In general bondy_router tries to handle all messages asynchronously. It does it by using either a static or a dynamic pool of workers based on configuration. This module implements both type of workers as a gen_server (this module). A static pool uses a set of supervised processes whereas a dynamic pool spawns a new erlang process for each message. In both cases, sidejob supervises the processes. By default bondy_router uses a dynamic pool.

The pools are implemented using the sidejob library in order to provide load regulation. Inn case a maximum pool capacity has been reached, the router will handle the message synchronously i.e. blocking the calling processes (usually the one that handles the transport connection e.g. bondy_wamp_ws_connection_handler).

The router also handles messages synchronously in those cases where it needs to preserve message ordering guarantees.

This module handles only the concurrency and basic routing logic, delegating the rest to either bondy_broker or bondy_dealer, which implement the actual PubSub and RPC logic respectively.

  ,------.                                    ,------.
  | Peer |                                    | Peer |
  `--+---'                                    `--+---'
     |                                           |
     |               TCP established             |
     |<----------------------------------------->|
     |                                           |
     |               TLS established             |
     |+<--------------------------------------->+|
     |+                                         +|
     |+           WebSocket established         +|
     |+|<------------------------------------->|+|
     |+|                                       |+|
     |+|            WAMP established           |+|
     |+|+<----------------------------------->+|+|
     |+|+                                     +|+|
     |+|+                                     +|+|
     |+|+            WAMP closed              +|+|
     |+|+<----------------------------------->+|+|
     |+|                                       |+|
     |+|                                       |+|
     |+|            WAMP established           |+|
     |+|+<----------------------------------->+|+|
     |+|+                                     +|+|
     |+|+                                     +|+|
     |+|+            WAMP closed              +|+|
     |+|+<----------------------------------->+|+|
     |+|                                       |+|
     |+|           WebSocket closed            |+|
     |+|<------------------------------------->|+|
     |+                                         +|
     |+              TLS closed                 +|
     |+<--------------------------------------->+|
     |                                           |
     |               TCP closed                  |
     |<----------------------------------------->|
     |                                           |
  ,--+---.                                    ,--+---.
  | Peer |                                    | Peer |
  `------'                                    `------'
(Diagram copied from WAMP RFC Draft)

Types

event/0

Specs

event() :: {wamp_message(), bondy_context:t()}.

id/0

Specs

id() :: 0..9007199254740992.

uri/0

Specs

uri() :: binary().

wamp_abort/0

Specs

wamp_abort() :: #abort{}.

wamp_authenticate/0

Specs

wamp_authenticate() :: #authenticate{}.

wamp_call/0

Specs

wamp_call() :: #call{}.

wamp_cancel/0

Specs

wamp_cancel() :: #cancel{}.

wamp_challenge/0

Specs

wamp_challenge() :: #challenge{}.

wamp_error/0

Specs

wamp_error() :: #error{}.

wamp_event/0

Specs

wamp_event() :: #event{}.

wamp_event_received/0

Specs

wamp_event_received() :: #event_received{}.

wamp_goodbye/0

Specs

wamp_goodbye() :: #goodbye{}.

wamp_hello/0

Specs

wamp_hello() :: #hello{}.

wamp_interrupt/0

Specs

wamp_interrupt() :: #interrupt{}.

wamp_invocation/0

Specs

wamp_invocation() :: #invocation{}.

wamp_message/0

Specs

wamp_message() ::
    wamp_hello() |
    wamp_challenge() |
    wamp_authenticate() |
    wamp_welcome() |
    wamp_abort() |
    wamp_goodbye() |
    wamp_error() |
    wamp_publish() |
    wamp_published() |
    wamp_subscribe() |
    wamp_subscribed() |
    wamp_unsubscribe() |
    wamp_unsubscribed() |
    wamp_event() |
    wamp_event_received() |
    wamp_subscriber_received() |
    wamp_call() |
    wamp_cancel() |
    wamp_result() |
    wamp_register() |
    wamp_registered() |
    wamp_unregister() |
    wamp_unregistered() |
    wamp_invocation() |
    wamp_interrupt() |
    wamp_yield().

wamp_publish/0

Specs

wamp_publish() :: #publish{}.

wamp_published/0

Specs

wamp_published() :: #published{}.

wamp_register/0

Specs

wamp_register() :: #register{}.

wamp_registered/0

Specs

wamp_registered() :: #registered{}.

wamp_result/0

Specs

wamp_result() :: #result{}.

wamp_subscribe/0

Specs

wamp_subscribe() :: #subscribe{}.

wamp_subscribed/0

Specs

wamp_subscribed() :: #subscribed{}.

wamp_subscriber_received/0

Specs

wamp_subscriber_received() :: #subscriber_received{}.

wamp_unregister/0

Specs

wamp_unregister() :: #unregister{}.

wamp_unregistered/0

Specs

wamp_unregistered() :: #unregistered{}.

wamp_unsubscribe/0

Specs

wamp_unsubscribe() :: #unsubscribe{}.

wamp_unsubscribed/0

Specs

wamp_unsubscribed() :: #unsubscribed{}.

wamp_welcome/0

Specs

wamp_welcome() :: #welcome{}.

wamp_yield/0

Specs

wamp_yield() :: #yield{}.

Functions

agent()

Returns the Bondy agent identification string

close_context(Ctxt)

Specs

close_context(bondy_context:t()) -> bondy_context:t().

forward(M, Ctxt)

Specs

forward(M :: wamp_message(), Ctxt :: bondy_context:t()) ->
           {ok, bondy_context:t()} |
           {reply, Reply :: wamp_message(), bondy_context:t()} |
           {stop, Reply :: wamp_message(), bondy_context:t()}.

Forwards a WAMP message to the Dealer or Broker based on message type. The message might end up being handled synchronously (performed by the calling process i.e. the transport handler) or asynchronously (by sending the message to the router load regulated worker pool).

forward(Msg, To, Opts)

Specs

forward(wamp_message(), bondy_ref:t(), map()) -> ok | no_return().

pre_stop()

Sends a GOODBYE message to all existing client connections. The client should reply with another GOODBYE within the configured time and when it does or on timeout, Bondy will close the connection triggering the cleanup of all the client sessions.

roles()

Specs

roles() -> #{binary() => #{binary() => boolean()}}.

stop()