bondy_rpc_load_balancer (bondy v1.0.0-beta.16)

This module implements a distributed load balancer, providing the different load balancing strategies used by bondy_dealer to choose the Callee and Procedure to invoke when handling a WAMP Call.

At the moment the load balancing state is local and not replicated across the nodes in the cluster. However, each node has access to a local replica of the global registry and thus can load balance between local and remote Callees.

## Supported Load Balancing Strategies

Bondy supports all WAMP Basic and Advanced Profile load balancing strategies for Shared Registrations and extends those with additional strategies.

### Single

### First

### Last

### Random

### Round Robin

### Jump Consistent Hash

### Queue Least Loaded

### Queue Least Loaded Sample
In the future we will explore implementing distributed load balancing algorithms such as Ant Colony, Particle Swarm Optimization and Biased Random Sampling [See references](https://pdfs.semanticscholar.org/b9a9/52ed1b8bfae2e976b5c0106e894bd4c41d89.pdf)

Types

entries/0

Specs

entries() :: [bondy_registry_entry:t()].

iterator/0
(opaque)

Specs

iterator()

opts/0

Specs

opts() :: #{strategy := strategy(), x_force_locality => boolean(), x_routing_key => binary()}.

strategy/0

Specs

strategy() ::
    single | first | last | random | round_robin | jump_consistent_hash | queue_least_loaded |
    queue_least_loaded_sample.

uri/0

Specs

uri() :: binary().

Functions

get(Entries, Opts)

Specs

get(entries(), opts()) -> bondy_registry_entry:t() | {error, noproc}.

iterate(Iterator)

Specs

iterate(iterator()) -> {bondy_registry_entry:t(), iterator()} | {error, noproc} | end_of_table.

iterate(Entries, Opts0)

Specs

iterate(entries(), opts()) ->
           {bondy_registry_entry:t(), iterator()} | '$end_of_table' | {error, noproc | map()}.