bondy_rpc_promise (bondy v1.0.0-beta.16)

A promise is used to implement a capability and a feature: - the capability to match the callee response (wamp_yield() or wamp_error()) back to the origin wamp_call() and Caller - the call_timeout feature at the dealer level

Types

dequeue_fun/0

Specs

dequeue_fun() :: fun((empty | {ok, t()}) -> any()).

id/0

Specs

id() :: 0..9007199254740992.

key/0

Specs

key() ::
    {RealmUri :: uri(),
     InvocationId :: wildcard(id()),
     CallId :: wildcard(id()),
     CalleeNodestring :: wildcard(nodestring()),
     CalleeSession :: wildcard(id()),
     CallerNodestring :: wildcard(nodestring()),
     CallerSession :: wildcard(id())}.

match_opts/0

Specs

match_opts() :: #{id => id(), caller => bondy_ref:t(), callee => bondy_ref:t()}.

nodestring/0

Specs

nodestring() :: binary().

opts/0

Specs

opts() ::
    #{call_id => id(),
      procedure_uri => uri(),
      via => bondy_ref:relay() | bondy_ref:bridge_relay()}.

t/0
(opaque)

Specs

t()

uri/0

Specs

uri() :: binary().

wildcard/1

Specs

wildcard(T) :: T | '_'.

Functions

call_id(Bondy_rpc_promise)

Returns the call request identifier

callee(Bondy_rpc_promise)

Specs

callee(t()) -> bondy_ref:t().

Returns the callee (bondy_ref:t()) that is the target of this invocation promise.

caller(Bondy_rpc_promise)

Specs

caller(t()) -> bondy_ref:t().

Returns the caller (bondy_ref:t()) who made the call request associated with this invocation promise.

dequeue(Key)

Specs

dequeue(Key :: key()) -> empty | {ok, t()}.

Dequeues the promise that matches key pattern

dequeue(Key, Fun)

Specs

dequeue(key(), dequeue_fun()) -> any().

Dequeues the promise that matches key pattern

enqueue(RealmUri, Bondy_rpc_promise, Timeout)

Adds the invocation promise P to the promise queue for realm RealmUri using a timeout of Timeout.
If the promise is not dequeued before Timeout milliseconds, the caller will receive an error with reason "wamp.error.timeout".

flush(Ref)

Specs

flush(bondy_ref:t()) -> ok.

Removes all pending promises from the queue for the reference

invocation_id(Bondy_rpc_promise)

Returns the invocation request identifier

key_pattern(RealmUri, InvocationId, CallId, Callee, Caller)

Specs

key_pattern(RealmUri :: uri(),
            InvocationId :: id(),
            CallId :: id(),
            Callee :: bondy_ref:t(),
            Caller :: bondy_ref:t()) ->
               key().

Pattern for dequeueing

new(InvocationId, Callee, Ctxt)

Specs

new(InvocationId :: id(), Callee :: bondy_ref:t(), Ctxt :: bondy_context:t()) -> t().

Creates a new promise for a remote invocation. In this case the caller lives in a different node and has recorded a promise using new/5 in its node.

new(InvocationId, Callee, Ctxt, Opts)

Specs

new(InvocationId :: id(), Callee :: bondy_ref:t(), Ctxt :: bondy_context:t(), Opts :: opts()) ->
       t().

Creates a new promise for a remote invocation. In this case the caller lives in a different node and has recorded a promise using new/5 in its node.

peek(Key)

Specs

peek(key()) -> {ok, t()} | empty.

Reads the promise that matches the key pattern

procedure_uri(Bondy_rpc_promise)

queue_size()

timestamp(Bondy_rpc_promise)

via(Bondy_rpc_promise)

Specs

via(t()) -> queue:queue(bondy_ref:relay() | bondy_ref:bridge_relay()).

Returns the queue of relays that are needed to forward an invocation result to the caller.