bondy_sensitive behaviour (bondy v1.0.0-beta.16)

Types

sensitive/0

Specs

sensitive() :: {sensitive, fun()}.

Callbacks

format_status/2

Specs

format_status(Opts :: normal | terminate, State :: term()) -> term().

Functions

conforms(Mod)

Specs

conforms(Mod :: module()) -> boolean().

Returns true is module Mod conforms with this behaviour.

format_status(Opt, Mod, State)

Specs

format_status(Opt :: normal | terminate, Mod :: module(), State :: term()) -> term().

Formalises and extends the use of the callback format_status/2 gen_server callback ([See documentation here](https://erlang.org/doc/man/gen_server.html#Module:format_status-2)) to modules.

gen_server:format_status/2 is called by a gen_server process in the following situations:

* One of sys:get_status/1,2 is invoked to get the gen_server status. Opt is set to the atom normal. * The gen_server process terminates abnormally and logs an error. Opt is set to the atom terminate.

A callback module managing the status for or on-behalf of a gen_server porocess can use this function to change the return value. State is the internal state of the gen_server process.
The function is to return Status, a term that changes the details of the current state State and status of the gen_server process. There are no restrictions on the form Status can take.

raise(Class, Reason, Stacktrace)

Specs

raise(Class :: error | exit | throw, Reason :: term(), Stacktrace :: erlang:raise_stacktrace()) ->
         badarg.

unwrap(_)

Specs

unwrap(sensitive()) -> term().

wrap(Term)

Specs

wrap(Term :: term() | fun(() -> term())) -> sensitive().