bondy_session (bondy v1.0.0-beta.16)

A Session (wamp session) is a transient conversation between two WAMP Peers attached to a Realm and running over a Transport.

Bondy implementation ties the lifetime of the underlying transport connection for a WAMP connection to that of a WAMP Session i.e. establish a new transport-layer connection as part of each new session establishment.

A Bondy Session is a not an application Session and is not a store for application specific content (an application session store should be implemented as a service i.e. a Callee).
Currently sessions are not persistent i.e. if the connection closes the session data will be lost.

Types

details/0

Specs

details() ::
    #{session => id(),
      authid => id(),
      authrole => binary(),
      authmethod => binary(),
      authprovider => binary(),
      transport => #{agent => binary(), peername => binary()}}.

id/0

Specs

id() :: 0..9007199254740992.

maybe/1

Specs

maybe(T) :: T | undefined.

nodestring/0

Specs

nodestring() :: binary().

peer/0

Specs

peer() :: {inet:ip_address(), inet:port_number()}.

properties/0

Specs

properties() ::
    #{roles := map(),
      agent => binary(),
      authid => binary(),
      authrole => binary(),
      authroles => [binary()],
      authmethod => binary(),
      is_anonymous => boolean(),
      roles => peer()}.

t/0

Specs

t() :: #session{}.

uri/0

Specs

uri() :: binary().

Functions

agent(Session)

Specs

agent(t()) -> binary() | undefined.

authid(Session)

Specs

authid(t()) -> bondy_rbac_user:username().

authmethod(Session)

Specs

authmethod(t()) -> binary().

authrole(Session)

Specs

authrole(t()) -> maybe(bondy_rbac_group:name()).

authroles(Session)

Specs

authroles(t()) -> [bondy_rbac_group:name()].

close(Session)

Specs

close(t()) -> ok.

created(Session)

Specs

created(t()) -> pos_integer().

Returns the time at which the session was created, Its value is a timestamp in seconds.

fetch(Id)

Specs

fetch(id()) -> t() | no_return().

Retrieves the session identified by Id from the tuplespace. If the session does not exist it fails with reason '{badarg, Id}'.

format_status(Opt, Session)

Specs

format_status(Opt :: normal | terminate, Session :: t()) -> term().

id(Session)

Specs

id(t()) -> id().

incr_seq(Session)

Specs

incr_seq(id() | t()) -> map().

info(Session)

Specs

info(t()) -> details().

is_security_enabled(Session)

Specs

is_security_enabled(t()) -> boolean().

list()

list(_)

list_refs(N)

list_refs(RealmUri, N)

lookup(Id)

Specs

lookup(id() | bondy_ref:t()) -> t() | {error, not_found}.

Retrieves the session identified by Id from the tuplespace or 'not_found' if it doesn't exist.

lookup(RealmUri, Id)

Specs

lookup(RealmUri :: uri(), Id :: id()) -> t() | {error, not_found}.

Retrieves the session identified by Id from the tuplespace or 'not_found' if it doesn't exist.

new(RealmUri, Opts)

Specs

new(uri() | bondy_realm:t(), properties()) -> t() | no_return().

Creates a new transient session (not persisted)

new(Id, RealmUri, Opts)

Specs

new(id(), uri() | bondy_realm:t(), properties()) -> t() | no_return().

node(Session)

Specs

node(t()) -> atom().

Returns the node of the process managing the transport that the session identified by Id runs on.

nodestring(Session)

Specs

nodestring(t()) -> nodestring().

Returns the node of the process managing the transport that the session identified by Id runs on.

peer(Session)

Specs

peer(t()) -> peer().

pid(Session)

Specs

pid(id() | t()) -> pid().

Returns the pid of the process managing the transport that the session identified by Id runs on.

rbac_context(Session)

Specs

rbac_context(id() | t()) -> bondy_rbac:context().

realm_uri(Session)

Specs

realm_uri(id() | t()) -> uri().

ref(Session)

Specs

ref(t()) -> bondy_ref:client() | bondy_ref:relay().

Returns the identifier for the owner of this session

refresh_rbac_context(Session)

Specs

refresh_rbac_context(id() | t()) -> bondy_rbac:context().

roles(Session)

Specs

roles(id() | t()) -> map().

size()

Specs

size() -> non_neg_integer().

Returns the number of sessions in the tuplespace.

store(Session)

Specs

store(t()) -> ok | no_return().

Creates a new session provided the RealmUri exists or can be dynamically created. It calls bondy_utils:get_realm/1 which will fail with an exception if the realm does not exist or cannot be created -----------------------------------------------------------------------------

to_external(Session)

Specs

to_external(t()) -> details().

type(Session)

Specs

type(t()) -> bondy_ref:type().

update(Session)

Specs

update(t()) -> ok.

user(Session)

Specs

user(t()) -> bondy_rbac_user:t().