bondy_session_manager (bondy v1.0.0-beta.16)

Types

uri/0

Specs

uri() :: binary().

Functions

close(Session)

Specs

close(bondy_session:t()) -> ok.

code_change(OldVsn, State, Extra)

handle_call(Event, From, State)

handle_cast(Event, State)

handle_info(Info, State)

init(_)

open(Id, RealmOrUri, Opts)

Specs

open(bondy_session:id(), uri() | bondy_realm:t(), bondy_session:properties()) ->
        bondy_session:t() | no_return().

Creates a new session provided the RealmUri exists or can be dynamically created. It calls bondy_session:new/4 which will fail with an exception if the realm does not exist or cannot be created.
This function also sets up a monitor for the calling process which is assummed to be the client connection process e.g. WAMP connection. In case the connection crashes it performs the cleanup of any session data that should not be retained. -----------------------------------------------------------------------------

pool()

Specs

pool() -> term().

pool_size()

Specs

pool_size() -> integer().

start_link(Pool, Name)

terminate(Reason, State)