bondy_wamp_event_manager (bondy v1.0.0-beta.16)

This module provides a bridge between WAMP events and OTP events.

Functions

add_callback(Fun)

Subscribe to a WAMP event with a callback function. The function needs to have two arguments representing the topic_uri and the wamp_event() that has been published.

add_handler(Handler, Args)

Adds an event handler. Calls gen_event:add_handler(?MODULE, Handler, Args). The handler will receive all WAMP events.

add_sup_callback(Fun)

Subscribe to a WAMP event with a supervised callback function. The function needs to have two arguments representing the topic_uri and the wamp_event() that has been published.

add_sup_handler(Handler, Args)

Adds a supervised event handler. Calls gen_event:add_sup_handler(?MODULE, Handler, Args). The handler will receive all WAMP events.

code_change(OldVsn, State, Extra)

handle_call(Event, State)

handle_event(_, State)

handle_info(Info, State)

init(_)

notify(Topic, Event)

Notifies all event handlers of the event

start_link()

terminate(Reason, State)