bondy_wamp_ws_connection_handler (bondy v1.0.0-beta.16)

A Cowboy WS handler.

Each WAMP message is transmitted as a separate WebSocket message (not WebSocket frame)

The WAMP protocol MUST BE negotiated during the WebSocket opening handshake between Peers using the WebSocket subprotocol negotiation mechanism.

WAMP uses the following WebSocket subprotocol identifiers for unbatched modes:

* "wamp.2.json" * "wamp.2.msgpack"

With "wamp.2.json", _all_ WebSocket messages MUST BE of type *text* (UTF8 encoded) and use the JSON message serialization.

With "wamp.2.msgpack", _all_ WebSocket messages MUST BE of type *binary* and use the MsgPack message serialization.
To avoid incompatibilities merely due to naming conflicts with WebSocket subprotocol identifiers, implementers SHOULD register identifiers for additional serialization formats with the official WebSocket subprotocol registry.

Types

state/0

Specs

state() :: #state{}.

Functions

init(Req0, _)

Specs

init(cowboy_req:req(), state()) ->
        {ok | module(), cowboy_req:req(), state()} |
        {module(), cowboy_req:req(), state(), hibernate} |
        {module(), cowboy_req:req(), state(), timeout()} |
        {module(), cowboy_req:req(), state(), timeout(), hibernate}.

terminate(Other, Req, St)

Termination

websocket_handle(Data, State)

Handles frames sent by client

websocket_info(_, St)

Handles internal erlang messages and WAMP messages BONDY wants to send to the client. See bondy:send/2.

websocket_init(State)

Initialises the WS connection.